{"schema":"decodex/commit/1","summary":"Modularize autonomy and intake surfaces","authority":"XY-1115"}
