+++ apps/decodex/src/autonomy_proposal.rs
@@ -16,6 +16,13 @@ use crate::{
 	prelude::{Result, eyre},
 };
 
+mod decision;
+#[cfg(test)] mod tests;
+mod validation;
+
+#[allow(clippy::wildcard_imports)] use decision::*;
+#[allow(clippy::wildcard_imports)] use validation::*;
+
 pub(crate) const AUTONOMY_PROPOSAL_SCHEMA: &str = "decodex.autonomy_proposal/1";
 
 const AUTONOMY_PROPOSAL_RECORD_VERSION: u16 = 1;
@@ -928,1690 +935,5 @@ impl AutonomyProposal {
 impl AutonomyProposal {
 	fn decision_contract_id(&self) -> String {
 		format!("autonomy-decision-{}", &self.fingerprint[..32])
-	}
-}
-
-fn autonomy_decision_research_provenance(
-	proposal: &AutonomyProposal,
-	authority: &AutonomyProposalDecisionBridgeAuthority,
-) -> Vec<Value> {
-	let mut provenance = vec![
-		serde_json::json!({
-			"kind": "autonomy_proposal",
-			"reference": proposal.id.clone(),
-			"summary": proposal.summary.clone(),
-		}),
-		serde_json::json!({
-			"kind": "autonomy_objective",
-			"reference": format!(
-				"{}:{}@{}",
-				proposal.project_id, proposal.objective_id, proposal.objective_version
-			),
-			"summary": proposal.objective_lineage.objective_summary.as_deref().unwrap_or(
-				"Accepted autonomy objective version."
-			),
-		}),
-		serde_json::json!({
-			"kind": "proposal_acceptance",
-			"reference": authority.acceptance_source.clone(),
-			"summary": format!(
-				"Accepted by {} ({}) at {}.",
-				authority.accepted_by,
-				authority.accepted_by_kind.as_str(),
-				authority.accepted_at
-			),
-		}),
-	];
-
-	if let Some(policy) = &authority.accepted_project_policy {
-		provenance.push(serde_json::json!({
-			"kind": "project_policy",
-			"reference": policy.authority_ref.clone(),
-			"summary": format!(
-				"Accepted project policy {}@{} authorized `{}` ({}) for autonomy proposal acceptance.",
-				policy.accepted_policy_id,
-				policy.accepted_policy_version,
-				policy.authorized_actor,
-				policy.authorized_actor_kind.as_str()
-			)
-		}));
-	}
-
-	provenance
-}
-
-fn autonomy_decision_research_evidence(proposal: &AutonomyProposal) -> Vec<Value> {
-	let mut evidence = proposal
-		.source_signals
-		.iter()
-		.map(|signal| {
-			let mut support = vec![
-				format!("freshness={}", signal.freshness),
-				format!("evidence_class={}", signal.evidence_class),
-				format!("confidence={}", signal.confidence),
-			];
-
-			if !signal.gaps.is_empty() {
-				support.push(format!("gaps={}", signal.gaps.join("; ")));
-			}
-			if !signal.contradictions.is_empty() {
-				support.push(format!("contradictions={}", signal.contradictions.join("; ")));
-			}
-
-			serde_json::json!({
-				"kind": format!("autonomy_signal:{}", signal.kind),
-				"claim": format!("Autonomy signal `{}` contributed to accepted proposal `{}`.", signal.signal_id, proposal.id),
-				"support": support.join("; "),
-				"source_ref": signal.signal_id.clone(),
-			})
-		})
-		.collect::<Vec<_>>();
-
-	if !proposal.gaps.is_empty() {
-		evidence.push(serde_json::json!({
-			"kind": "autonomy_proposal_gap",
-			"claim": "Accepted proposal retained evidence gaps for downstream review.",
-			"support": proposal.gaps.join("; "),
-			"source_ref": proposal.id.clone(),
-		}));
-	}
-	if !proposal.contradictions.is_empty() {
-		evidence.push(serde_json::json!({
-			"kind": "autonomy_proposal_contradiction",
-			"claim": "Accepted proposal retained contradictions for downstream authority checks.",
-			"support": proposal.contradictions.join("; "),
-			"source_ref": proposal.id.clone(),
-		}));
-	}
-
-	for challenge in &proposal.challenge_evidence {
-		evidence.push(serde_json::json!({
-			"kind": "autonomy_proposal_challenge",
-			"claim": challenge.summary.clone(),
-			"support": challenge_support(challenge),
-			"source_ref": format!("challenge:{}", challenge.actor),
-		}));
-	}
-
-	evidence
-}
-
-fn challenge_support(challenge: &AutonomyProposalChallengeEvidence) -> String {
-	if !challenge.objections.is_empty() {
-		challenge.objections.join("; ")
-	} else if !challenge.evidence_refs.is_empty() {
-		format!("evidence_refs={}", challenge.evidence_refs.join("; "))
-	} else {
-		String::from("Challenge recorded no objections.")
-	}
-}
-
-fn autonomy_decision_research_options(proposal: &AutonomyProposal) -> Vec<Value> {
-	let mut options = vec![serde_json::json!({
-		"option": proposal.summary.clone(),
-		"tradeoffs": option_tradeoffs(proposal),
-		"decision": "accepted_as_latent_decision_contract",
-		"rejected_reason": null,
-	})];
-
-	options.extend(proposal.rejected_alternatives.iter().map(|alternative| {
-		serde_json::json!({
-			"option": alternative.clone(),
-			"tradeoffs": [],
-			"decision": null,
-			"rejected_reason": "Rejected before autonomy proposal acceptance.",
-		})
-	}));
-
-	options
-}
-
-fn option_tradeoffs(proposal: &AutonomyProposal) -> Vec<String> {
-	let mut tradeoffs = Vec::new();
-
-	tradeoffs.push(format!("Rollback path: {}", proposal.rollback_path));
-	tradeoffs.extend(
-		proposal.review_requirements.iter().map(|item| format!("Review requirement: {item}")),
-	);
-	tradeoffs
-		.extend(proposal.validation_gates.iter().map(|item| format!("Validation gate: {item}")));
-
-	tradeoffs
-}
-
-fn proposal_objectives(proposal: &AutonomyProposal) -> Vec<String> {
-	if proposal.goals.is_empty() { vec![proposal.summary.clone()] } else { proposal.goals.clone() }
-}
-
-fn proposal_constraints(proposal: &AutonomyProposal) -> Vec<String> {
-	let mut constraints = proposal
-		.allowed_surfaces
-		.iter()
-		.map(|surface| format!("Allowed surface: {surface}"))
-		.collect::<Vec<_>>();
-
-	constraints.extend(
-		proposal
-			.review_requirements
-			.iter()
-			.map(|requirement| format!("Review requirement: {requirement}")),
-	);
-	constraints.extend(
-		proposal
-			.challenge_requirements
-			.iter()
-			.map(|requirement| format!("Challenge requirement: {requirement}")),
-	);
-
-	for challenge in &proposal.challenge_evidence {
-		constraints.extend(
-			challenge
-				.objections
-				.iter()
-				.map(|objection| format!("Challenge promotion constraint: {objection}")),
-		);
-	}
-
-	constraints.push(String::from(
-		"Accepted autonomy proposal must remain latent until Decision Contract promotion.",
-	));
-
-	constraints
-}
-
-fn proposal_assumptions(
-	proposal: &AutonomyProposal,
-	authority: &AutonomyProposalDecisionBridgeAuthority,
-) -> Vec<String> {
-	let mut assumptions =
-		proposal.metrics.iter().map(|metric| format!("Metric: {metric}")).collect::<Vec<_>>();
-
-	assumptions.push(format!(
-		"Proposal actor `{}` ({}) is distinct from acceptance authority `{}` ({}) unless resolved accepted project policy is recorded.",
-		authority.proposal_actor,
-		authority.proposal_actor_kind.as_str(),
-		authority.accepted_by,
-		authority.accepted_by_kind.as_str()
-	));
-
-	assumptions
-}
-
-fn proposal_objections(proposal: &AutonomyProposal) -> Vec<String> {
-	let mut objections = proposal.contradictions.clone();
-
-	objections.extend(proposal.gaps.iter().map(|gap| format!("Evidence gap: {gap}")));
-
-	for challenge in &proposal.challenge_evidence {
-		objections.extend(challenge.objections.iter().cloned());
-	}
-
-	unique_sorted_strings(objections)
-}
-
-fn proposal_stop_conditions(proposal: &AutonomyProposal) -> Vec<String> {
-	let mut stop_conditions = vec![
-		String::from(
-			"Stop before Program Intake unless the Decision Contract is promoted with explicit authority.",
-		),
-		format!("Rollback path: {}", proposal.rollback_path),
-	];
-
-	stop_conditions.extend(
-		proposal
-			.challenge_requirements
-			.iter()
-			.map(|requirement| format!("Challenge requirement: {requirement}")),
-	);
-
-	stop_conditions
-}
-
-fn proposal_validation_expectations(proposal: &AutonomyProposal) -> Vec<String> {
-	if proposal.validation_gates.is_empty() {
-		vec![String::from("Run the accepted Decision Contract validation gate before promotion.")]
-	} else {
-		proposal.validation_gates.clone()
-	}
-}
-
-fn proposal_risk_notes(proposal: &AutonomyProposal) -> Vec<String> {
-	let mut risk_notes =
-		proposal.gaps.iter().map(|gap| format!("Evidence gap: {gap}")).collect::<Vec<_>>();
-
-	risk_notes.extend(
-		proposal
-			.review_requirements
-			.iter()
-			.map(|requirement| format!("Review requirement: {requirement}")),
-	);
-	risk_notes.extend(
-		proposal
-			.challenge_requirements
-			.iter()
-			.map(|requirement| format!("Challenge requirement: {requirement}")),
-	);
-
-	risk_notes
-}
-
-fn proposal_issue_candidate(proposal: &AutonomyProposal) -> Value {
-	serde_json::json!({
-		"key": format!("autonomy-{}", stable_slug(&proposal.source_family, 48)),
-		"title": proposal.summary.clone(),
-		"objective": proposal.summary.clone(),
-		"stage": proposal_issue_stage(&proposal.intended_surface),
-		"dependencies": [],
-		"conflict_domains": proposal_conflict_domains(proposal),
-		"acceptance": proposal_objectives(proposal),
-		"validation": proposal_validation_expectations(proposal),
-		"risk": proposal_risk_notes(proposal),
-		"queue_intent": "ready_to_queue",
-	})
-}
-
-fn proposal_conflict_domains(proposal: &AutonomyProposal) -> Vec<String> {
-	let mut domains = BTreeSet::new();
-
-	if let Some(surface) = normalize_repo_relative_path(&proposal.intended_surface) {
-		domains.insert(format!("file:{surface}"));
-	} else {
-		domains.insert(format!("surface:{}", proposal.intended_surface));
-	}
-
-	for identifier in &proposal.affected_identifiers {
-		domains.insert(format!("identifier:{identifier}"));
-	}
-
-	domains.into_iter().collect()
-}
-
-fn proposal_issue_stage(intended_surface: &str) -> &'static str {
-	if intended_surface.starts_with("docs/spec/") {
-		"spec"
-	} else if intended_surface.starts_with("docs/") {
-		"design"
-	} else if intended_surface.starts_with("apps/decodex/src/") {
-		"runtime"
-	} else {
-		"design"
-	}
-}
-
-fn proposal_source_issue_identifier(affected_identifiers: &[String]) -> Option<String> {
-	affected_identifiers
-		.iter()
-		.find(|identifier| looks_like_tracker_issue_identifier(identifier))
-		.cloned()
-}
-
-fn looks_like_tracker_issue_identifier(value: &str) -> bool {
-	let Some((prefix, number)) = value.split_once('-') else {
-		return false;
-	};
-
-	!prefix.is_empty()
-		&& prefix.bytes().all(|byte| byte.is_ascii_uppercase())
-		&& !number.is_empty()
-		&& number.bytes().all(|byte| byte.is_ascii_digit())
-}
-
-fn stable_slug(value: &str, max_len: usize) -> String {
-	let mut slug = String::new();
-	let mut previous_dash = false;
-
-	for character in value.chars() {
-		if character.is_ascii_alphanumeric() {
-			slug.push(character.to_ascii_lowercase());
-
-			previous_dash = false;
-		} else if !previous_dash && !slug.is_empty() {
-			slug.push('-');
-
-			previous_dash = true;
-		}
-		if slug.len() >= max_len {
-			break;
-		}
-	}
-
-	while slug.ends_with('-') {
-		slug.pop();
-	}
-
-	if slug.is_empty() { String::from("proposal") } else { slug }
-}
-
-fn proposal_refusals(
-	objective: Option<&AutonomyObjectiveContract>,
-	signals: &[AutonomySignal],
-	input: &AutonomyProposalCompileInput,
-	contradictions: &[String],
-) -> Vec<AutonomyProposalRefusal> {
-	let mut refusals = Vec::new();
-
-	match objective {
-		Some(objective)
-			if objective.project_id() == input.project_id
-				&& objective.id() == input.objective_id
-				&& objective.version() == input.objective_version
-				&& objective.state() == AutonomyObjectiveState::Accepted => {
-				for signal in signals {
-					if signal.project_id() != input.project_id
-						|| signal.objective_id() != input.objective_id
-						|| signal.objective_version() != input.objective_version
-					{
-						refusals.push(AutonomyProposalRefusal::new(
-							AutonomyProposalRefusalReason::MissingObjective,
-							format!(
-								"Signal `{}` is not tied to objective `{}` version {}.",
-								signal.id(),
-								input.objective_id,
-								input.objective_version
-							),
-							vec![signal.id().to_owned()],
-						));
-					}
-					if !objective
-						.allowed_signal_kinds()
-						.iter()
-						.any(|kind| kind == signal.kind().as_str())
-					{
-						refusals.push(AutonomyProposalRefusal::new(
-							AutonomyProposalRefusalReason::DisallowedSignalKind,
-							format!(
-								"Signal `{}` kind `{}` is outside the accepted objective allowed_signal_kinds.",
-								signal.id(),
-								signal.kind().as_str()
-							),
-							vec![signal.id().to_owned()],
-						));
-					}
-					if signal.freshness() != AutonomySignalFreshness::Fresh {
-						refusals.push(AutonomyProposalRefusal::new(
-							AutonomyProposalRefusalReason::StaleEvidence,
-							format!(
-								"Signal `{}` freshness is `{}` and requires fresh readback before acceptance.",
-								signal.id(),
-								signal.freshness().as_str()
-							),
-							vec![signal.id().to_owned()],
-						));
-					}
-				}
-
-				if !surface_allowed(&input.intended_surface, objective.allowed_surfaces()) {
-					refusals.push(AutonomyProposalRefusal::new(
-						AutonomyProposalRefusalReason::DisallowedSurface,
-						format!(
-							"Intended surface `{}` is outside the accepted objective allowed_surfaces.",
-							input.intended_surface
-						),
-						vec![input.objective_id.clone()],
-					));
-				}
-			},
-		Some(objective) => refusals.push(AutonomyProposalRefusal::new(
-			AutonomyProposalRefusalReason::MissingObjective,
-			format!(
-				"Objective `{}` version {} exists in state `{}` but is not the accepted exact proposal objective.",
-				objective.id(),
-				objective.version(),
-				objective.state().as_str()
-			),
-			vec![input.objective_id.clone()],
-		)),
-		None => refusals.push(AutonomyProposalRefusal::new(
-			AutonomyProposalRefusalReason::MissingObjective,
-			format!(
-				"Objective `{}` version {} is missing.",
-				input.objective_id,
-				input.objective_version
-			),
-			vec![input.objective_id.clone()],
-		)),
-	}
-
-	for contradiction in contradictions {
-		refusals.push(AutonomyProposalRefusal::new(
-			AutonomyProposalRefusalReason::UnresolvedContradiction,
-			format!("Contradiction remains unresolved: {contradiction}"),
-			vec![input.objective_id.clone()],
-		));
-	}
-	for note in &input.weakened_validation_or_review {
-		refusals.push(AutonomyProposalRefusal::new(
-			AutonomyProposalRefusalReason::WeakenedValidationReview,
-			format!("Validation or review evidence is weakened: {note}"),
-			vec![input.objective_id.clone()],
-		));
-	}
-
-	refusals
-}
-
-fn derive_proposal_state(
-	has_signals: bool,
-	refusals: &[AutonomyProposalRefusal],
-) -> AutonomyProposalState {
-	if refusals.iter().any(|refusal| {
-		matches!(
-			refusal.reason,
-			AutonomyProposalRefusalReason::DisallowedSignalKind
-				| AutonomyProposalRefusalReason::DisallowedSurface
-		)
-	}) {
-		return AutonomyProposalState::Rejected;
-	}
-	if refusals
-		.iter()
-		.any(|refusal| refusal.reason == AutonomyProposalRefusalReason::UnresolvedContradiction)
-	{
-		return AutonomyProposalState::NeedsHumanDecision;
-	}
-	if !refusals.is_empty() {
-		return AutonomyProposalState::NeedsEvidence;
-	}
-	if has_signals {
-		AutonomyProposalState::DecisionCandidate
-	} else {
-		AutonomyProposalState::Draft
-	}
-}
-
-fn surface_allowed(intended_surface: &str, allowed_surfaces: &[String]) -> bool {
-	let Some(intended_surface) = normalize_repo_relative_path(intended_surface) else {
-		return false;
-	};
-
-	allowed_surfaces.iter().any(|surface| {
-		normalize_repo_relative_path(surface).is_some_and(|surface| {
-			intended_surface == surface
-				|| intended_surface
-					.strip_prefix(&surface)
-					.is_some_and(|suffix| suffix.starts_with('/'))
-		})
-	})
-}
-
-fn normalize_repo_relative_path(value: &str) -> Option<String> {
-	let path = Path::new(value);
-
-	if path.is_absolute() {
-		return None;
-	}
-
-	let mut parts = Vec::new();
-
-	for component in path.components() {
-		let Component::Normal(part) = component else {
-			return None;
-		};
-
-		parts.push(part.to_str()?);
-	}
-
-	if parts.is_empty() {
-		return None;
-	}
-
-	Some(parts.join("/"))
-}
-
-fn autonomy_proposal_schema() -> String {
-	AUTONOMY_PROPOSAL_SCHEMA.to_owned()
-}
-
-const fn autonomy_proposal_record_version() -> u16 {
-	AUTONOMY_PROPOSAL_RECORD_VERSION
-}
-
-fn autonomy_proposal_id(fingerprint: &str) -> String {
-	format!("autonomy_proposal:{fingerprint}")
-}
-
-fn autonomy_proposal_fingerprint(proposal: &AutonomyProposal) -> Result<String> {
-	let material = serde_json::json!({
-		"project_id": proposal.project_id,
-		"objective_id": proposal.objective_id,
-		"objective_version": proposal.objective_version,
-		"source_signal_ids": proposal.source_signal_ids,
-		"affected_identifiers": proposal.affected_identifiers,
-		"source_family": proposal.source_family,
-		"intended_surface": proposal.intended_surface,
-	});
-	let payload = serde_json::to_vec(&material)?;
-	let digest = Sha256::digest(payload);
-	let mut hash = String::with_capacity(64);
-
-	for byte in digest {
-		hash.push(char::from(b"0123456789abcdef"[(byte >> 4) as usize]));
-		hash.push(char::from(b"0123456789abcdef"[(byte & 0x0f) as usize]));
-	}
-
-	Ok(hash)
-}
-
-fn validate_compile_input(input: &AutonomyProposalCompileInput) -> Result<()> {
-	validate_required("autonomy proposal input.project_id", &input.project_id)?;
-	validate_required("autonomy proposal input.objective_id", &input.objective_id)?;
-	validate_required("autonomy proposal input.source_family", &input.source_family)?;
-	validate_required("autonomy proposal input.intended_surface", &input.intended_surface)?;
-	validate_required("autonomy proposal input.summary", &input.summary)?;
-	validate_required("autonomy proposal input.rollback_path", &input.rollback_path)?;
-	validate_required("autonomy proposal input.created_at", &input.created_at)?;
-	validate_string_list(
-		"autonomy proposal input.affected_identifiers",
-		&input.affected_identifiers,
-	)?;
-	validate_string_list(
-		"autonomy proposal input.challenge_requirements",
-		&input.challenge_requirements,
-	)?;
-	validate_string_list(
-		"autonomy proposal input.rejected_alternatives",
-		&input.rejected_alternatives,
-	)?;
-	validate_string_list(
-		"autonomy proposal input.weakened_validation_or_review",
-		&input.weakened_validation_or_review,
-	)?;
-
-	if input.objective_version == 0 {
-		eyre::bail!("Autonomy proposal input objective_version must be greater than zero.");
-	}
-
-	Ok(())
-}
-
-fn validate_required(name: &str, value: &str) -> Result<()> {
-	if value.trim().is_empty() {
-		eyre::bail!("{name} must not be empty.");
-	}
-
-	Ok(())
-}
-
-fn validate_optional_required(name: &str, value: Option<&str>) -> Result<()> {
-	if let Some(value) = value {
-		validate_required(name, value)?;
-	}
-
-	Ok(())
-}
-
-fn validate_string_list(name: &str, values: &[String]) -> Result<()> {
-	for value in values {
-		validate_required(name, value)?;
-	}
-
-	Ok(())
-}
-
-fn validate_sorted_unique(name: &str, values: &[String]) -> Result<()> {
-	validate_string_list(name, values)?;
-
-	let mut seen = BTreeSet::new();
-	let mut previous = None;
-
-	for value in values {
-		if previous.is_some_and(|previous| previous > value.as_str()) {
-			eyre::bail!("{name} must be sorted.");
-		}
-		if !seen.insert(value.as_str()) {
-			eyre::bail!("{name} must not contain duplicates.");
-		}
-
-		previous = Some(value.as_str());
-	}
-
-	Ok(())
-}
-
-fn unique_sorted_strings(values: impl IntoIterator<Item = String>) -> Vec<String> {
-	values
-		.into_iter()
-		.map(|value| value.trim().to_owned())
-		.filter(|value| !value.is_empty())
-		.collect::<BTreeSet<_>>()
-		.into_iter()
-		.collect()
-}
-
-#[cfg(test)]
-mod tests {
-	use std::{collections::BTreeMap, slice};
-
-	use crate::{
-		autonomy_objective::{
-			AutonomyObjectiveAcceptance, AutonomyObjectiveActorKind, AutonomyObjectiveContract,
-		},
-		autonomy_proposal::{
-			AutonomyProposal, AutonomyProposalAcceptedProjectPolicy,
-			AutonomyProposalAuthorityActorKind, AutonomyProposalChallengeInput,
-			AutonomyProposalChallengeSource, AutonomyProposalCompileInput,
-			AutonomyProposalDecisionBridgeAuthority, AutonomyProposalRefusalReason,
-			AutonomyProposalState,
-		},
-		autonomy_signal::{
-			AutonomySignal, AutonomySignalConfidence, AutonomySignalEvidenceClass,
-			AutonomySignalFreshness, AutonomySignalInput, AutonomySignalPrivacy,
-			AutonomySignalSourceType,
-		},
-		loop_contract::{DecisionContractStatus, DecisionPromotion, DecisionPromotionActorKind},
-		state::{DecisionContractRecord, StateStore},
-	};
-
-	trait ExpectNone {
-		fn expect_none(self, message: &str);
-	}
-
-	impl<T> ExpectNone for Option<T> {
-		fn expect_none(self, message: &str) {
-			assert!(self.is_none(), "{message}");
-		}
-	}
-
-	fn objective_draft_fixture() -> AutonomyObjectiveContract {
-		serde_json::from_value(serde_json::json!({
-			"schema": "decodex.autonomy_objective/1",
-			"record_version": 1,
-			"project_id": "decodex",
-			"id": "quality-autonomy",
-			"version": 1,
-			"state": "draft",
-			"summary": "Improve Decodex autonomy quality under explicit authority.",
-			"goals": ["Reduce repeated validation and review churn."],
-			"non_goals": ["Do not bypass Decision Contract authority."],
-			"metrics": ["Validation retry count stays below objective tolerance."],
-			"allowed_surfaces": ["apps/decodex/src", "docs/spec"],
-			"allowed_signal_kinds": ["runtime_health", "review_feedback_cluster"],
-			"validation_gates": ["cargo test -p decodex autonomy_proposal --lib"],
-			"review_policy": "independent current-head review required",
-			"memory_policy": "read-only source-linked memory only",
-			"report_policy": "public-safe summaries only"
-		}))
-		.expect("draft objective should parse")
-	}
-
-	fn objective_fixture() -> AutonomyObjectiveContract {
-		let mut objective = objective_draft_fixture();
-
-		objective
-			.accept(
-				AutonomyObjectiveAcceptance::new(
-					"operator",
-					AutonomyObjectiveActorKind::User,
-					"2026-06-22T00:00:00Z",
-					"conversation",
-				)
-				.expect("acceptance should validate"),
-			)
-			.expect("objective should accept");
-
-		objective
-	}
-
-	fn store_accepted_objective(store: &StateStore) -> AutonomyObjectiveContract {
-		store
-			.upsert_autonomy_objective_draft("decodex", objective_draft_fixture())
-			.expect("objective should store");
-
-		store
-			.accept_autonomy_objective_version(
-				"decodex",
-				"quality-autonomy",
-				1,
-				AutonomyObjectiveAcceptance::new(
-					"operator",
-					AutonomyObjectiveActorKind::User,
-					"2026-06-22T00:00:00Z",
-					"conversation",
-				)
-				.expect("acceptance should validate"),
-			)
-			.expect("objective should accept")
-			.objective()
-			.clone()
-	}
-
-	fn signal_input() -> AutonomySignalInput {
-		AutonomySignalInput {
-			project_id: String::from("decodex"),
-			objective_id: String::from("quality-autonomy"),
-			objective_version: 1,
-			source_type: AutonomySignalSourceType::Runtime,
-			source_refs: vec![String::from("status:runtime-health")],
-			primary_source_refs: Vec::new(),
-			issue_id: Some(String::from("XY-1086")),
-			run_id: Some(String::from("xy-1086-attempt-1")),
-			attempt_id: Some(String::from("1")),
-			head_sha: Some(String::from("3cd19609c44cb18bff9e7a34a2f4853754afcee0")),
-			captured_at: String::from("2026-06-22T00:00:00Z"),
-			freshness: AutonomySignalFreshness::Fresh,
-			summary: String::from("Runtime status readback showed repeated friction."),
-			evidence: vec![String::from("status readback retained the repeated friction signal")],
-			evidence_class: AutonomySignalEvidenceClass::LiveReadback,
-			contradictions: Vec::new(),
-			gaps: vec![String::from("No dashboard comparison included.")],
-			confidence: AutonomySignalConfidence::Medium,
-			privacy: AutonomySignalPrivacy::Team,
-			observed_counts: BTreeMap::new(),
-			review_evidence: None,
-			proposal_only: true,
-			created_at: String::from("2026-06-22T00:00:05Z"),
-		}
-	}
-
-	fn runtime_signal() -> AutonomySignal {
-		AutonomySignal::runtime_health(signal_input()).expect("runtime signal should validate")
-	}
-
-	fn compile_input() -> AutonomyProposalCompileInput {
-		AutonomyProposalCompileInput {
-			project_id: String::from("decodex"),
-			objective_id: String::from("quality-autonomy"),
-			objective_version: 1,
-			source_family: String::from("runtime_status"),
-			intended_surface: String::from("apps/decodex/src/orchestrator/status.rs"),
-			affected_identifiers: vec![
-				String::from("OperatorLoopStatus"),
-				String::from("operator_status"),
-			],
-			summary: String::from("Compile a bounded proposal from runtime friction evidence."),
-			challenge_requirements: vec![String::from(
-				"Subagent or inline skeptic objections are evidence only.",
-			)],
-			rejected_alternatives: vec![String::from("Direct Decision Contract promotion.")],
-			rollback_path: String::from("Discard the dry-run proposal record."),
-			weakened_validation_or_review: Vec::new(),
-			created_at: String::from("2026-06-22T00:01:00Z"),
-		}
-	}
-
-	fn bridge_authority() -> AutonomyProposalDecisionBridgeAuthority {
-		AutonomyProposalDecisionBridgeAuthority::new(
-			"operator",
-			AutonomyProposalAuthorityActorKind::User,
-			"2026-06-22T00:03:00Z",
-			"conversation",
-			"Operator accepted the proposal for Decision Contract promotion.",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			None,
-		)
-		.expect("bridge authority should validate")
-	}
-
-	fn accepted_project_policy(
-		authorized_actor: &str,
-		authorized_actor_kind: AutonomyProposalAuthorityActorKind,
-		acceptance_source: &str,
-	) -> AutonomyProposalAcceptedProjectPolicy {
-		AutonomyProposalAcceptedProjectPolicy::new(
-			"decodex",
-			"quality-autonomy",
-			1,
-			"quality-autonomy-policy",
-			"1",
-			"decodex.runtime_policy:quality-autonomy-policy@1",
-			authorized_actor,
-			authorized_actor_kind,
-			vec![String::from(acceptance_source)],
-			vec![String::from(super::AUTONOMY_PROPOSAL_ACCEPTANCE_SCOPE)],
-		)
-		.expect("accepted project policy should validate")
-	}
-
-	fn runtime_policy_bridge_authority() -> AutonomyProposalDecisionBridgeAuthority {
-		AutonomyProposalDecisionBridgeAuthority::new(
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			"2026-06-22T00:03:00Z",
-			"runtime-policy",
-			"Accepted project policy allows this agent to accept the proposal.",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			Some(accepted_project_policy(
-				"subagent",
-				AutonomyProposalAuthorityActorKind::ExternalAgent,
-				"runtime-policy",
-			)),
-		)
-		.expect("policy-backed bridge authority should validate")
-	}
-
-	fn store_challenged_autonomy_candidate() -> (StateStore, String, DecisionContractRecord) {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let objective = store_accepted_objective(&store);
-		let signal = store
-			.record_autonomy_signal("decodex", runtime_signal())
-			.expect("signal should store")
-			.signal()
-			.clone();
-		let mut input = compile_input();
-
-		input.affected_identifiers.push(String::from("XY-1087"));
-
-		let mut proposal = AutonomyProposal::compile_dry_run(Some(&objective), &[signal], input)
-			.expect("proposal should compile");
-		let proposal_id = proposal.id().to_owned();
-
-		proposal
-			.record_challenge(AutonomyProposalChallengeInput {
-				source: AutonomyProposalChallengeSource::InlineSkeptic,
-				actor: String::from("inline"),
-				summary: String::from("Inline skeptic found no blocker to latent conversion."),
-				objections: Vec::new(),
-				evidence_refs: vec![String::from("challenge:inline")],
-				recorded_at: String::from("2026-06-22T00:02:00Z"),
-			})
-			.expect("no-objection challenge should preserve candidate state");
-
-		assert_eq!(proposal.state(), AutonomyProposalState::DecisionCandidate);
-
-		store.record_autonomy_proposal("decodex", proposal).expect("proposal should persist");
-
-		let candidate = store
-			.accept_autonomy_proposal_as_decision_contract_candidate(
-				"decodex",
-				&proposal_id,
-				bridge_authority(),
-			)
-			.expect("accepted proposal should become a latent Decision Contract");
-
-		(store, proposal_id, candidate)
-	}
-
-	fn assert_autonomy_candidate_shape(store: &StateStore, candidate: &DecisionContractRecord) {
-		assert_eq!(candidate.status(), DecisionContractStatus::DraftLatent);
-		assert!(candidate.contract().promotion().is_none());
-		assert_eq!(candidate.source_issue_id(), Some("XY-1087"));
-		assert_eq!(candidate.contract().source_intent().source_issue_identifier(), Some("XY-1087"));
-		assert!(
-			candidate
-				.contract()
-				.accepted_authority()
-				.accepted_objectives()
-				.contains(&String::from("Reduce repeated validation and review churn."))
-		);
-		assert!(candidate.contract().accepted_authority().constraints().contains(&String::from(
-			"Review requirement: independent current-head review required"
-		)));
-		assert_eq!(
-			candidate.contract().execution_readiness().validation_expectations(),
-			&[String::from("cargo test -p decodex autonomy_proposal --lib")]
-		);
-		assert!(
-			candidate
-				.contract()
-				.execution_readiness()
-				.risk_notes()
-				.contains(&String::from("Evidence gap: No dashboard comparison included."))
-		);
-		assert_eq!(candidate.contract().execution_readiness().proposed_issues().len(), 1);
-		assert!(
-			candidate.contract().execution_readiness().proposed_issues()[0]
-				.conflict_domains()
-				.contains(&String::from("file:apps/decodex/src/orchestrator/status.rs"))
-		);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-		assert!(store.list_program_intake_plans("decodex").expect("intake plans").is_empty());
-	}
-
-	#[test]
-	fn autonomy_proposal_challenge_source_accepts_legacy_support_agent_alias() {
-		let source: AutonomyProposalChallengeSource =
-			serde_json::from_value(serde_json::json!("subagent"))
-				.expect("canonical subagent source should parse");
-		let legacy_source: AutonomyProposalChallengeSource =
-			serde_json::from_value(serde_json::json!("support_agent"))
-				.expect("legacy support_agent source should parse");
-
-		assert_eq!(source, AutonomyProposalChallengeSource::Subagent);
-		assert!(
-			legacy_source == AutonomyProposalChallengeSource::Subagent,
-			"legacy support_agent should canonicalize to Subagent"
-		);
-		assert_eq!(
-			serde_json::to_value(legacy_source).expect("source should serialize"),
-			serde_json::json!("subagent")
-		);
-	}
-
-	#[test]
-	fn autonomy_proposal_dry_run_candidate_shows_lineage_signals_gates_and_gaps() {
-		let objective = objective_fixture();
-		let signal = runtime_signal();
-		let proposal =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[signal], compile_input())
-				.expect("proposal should compile");
-
-		assert_eq!(proposal.state(), AutonomyProposalState::DecisionCandidate);
-		assert_eq!(proposal.objective_id(), "quality-autonomy");
-		assert_eq!(proposal.objective_version(), 1);
-		assert_eq!(proposal.allowed_surfaces(), ["apps/decodex/src", "docs/spec"]);
-		assert_eq!(proposal.validation_gates(), ["cargo test -p decodex autonomy_proposal --lib"]);
-		assert_eq!(proposal.source_signal_ids().len(), 1);
-		assert_eq!(proposal.gaps(), ["No dashboard comparison included."]);
-		assert!(proposal.contradictions().is_empty());
-		assert!(proposal.refusal_reasons().is_empty());
-
-		let dry_run_json = serde_json::to_value(&proposal).expect("proposal should encode");
-
-		assert_eq!(dry_run_json["dry_run"], true);
-		assert_eq!(dry_run_json["non_executable"], true);
-		assert_eq!(dry_run_json["objective_lineage"]["objective_id"], "quality-autonomy");
-		assert_eq!(dry_run_json["source_signals"][0]["signal_id"], proposal.source_signal_ids()[0]);
-		assert_eq!(dry_run_json["allowed_surfaces"][0], "apps/decodex/src");
-		assert_eq!(dry_run_json["goals"][0], "Reduce repeated validation and review churn.");
-		assert_eq!(
-			dry_run_json["metrics"][0],
-			"Validation retry count stays below objective tolerance."
-		);
-		assert_eq!(dry_run_json["non_goals"][0], "Do not bypass Decision Contract authority.");
-		assert_eq!(
-			dry_run_json["review_requirements"][0],
-			"independent current-head review required"
-		);
-		assert_eq!(
-			dry_run_json["challenge_requirements"][0],
-			"Subagent or inline skeptic objections are evidence only."
-		);
-		assert_eq!(dry_run_json["rejected_alternatives"][0], "Direct Decision Contract promotion.");
-		assert_eq!(dry_run_json["rollback_path"], "Discard the dry-run proposal record.");
-		assert_eq!(
-			dry_run_json["validation_gates"][0],
-			"cargo test -p decodex autonomy_proposal --lib"
-		);
-		assert!(dry_run_json["refusal_reasons"].as_array().expect("refusals array").is_empty());
-	}
-
-	#[test]
-	fn autonomy_decision_bridge_accepts_candidate_as_latent_contract_with_lineage_readback() {
-		let (store, proposal_id, candidate) = store_challenged_autonomy_candidate();
-
-		assert_autonomy_candidate_shape(&store, &candidate);
-
-		let readback = store
-			.decision_contract("decodex", candidate.contract_id())
-			.expect("contract readback should work")
-			.expect("candidate should persist");
-
-		assert_eq!(readback.contract(), candidate.contract());
-
-		let idempotent = store
-			.accept_autonomy_proposal_as_decision_contract_candidate(
-				"decodex",
-				&proposal_id,
-				bridge_authority(),
-			)
-			.expect("re-accepting the same latent contract should be idempotent");
-
-		assert_eq!(idempotent.contract(), candidate.contract());
-
-		let missing_promotion_authority = DecisionPromotion::new(
-			"",
-			DecisionPromotionActorKind::User,
-			"2026-06-22T00:04:00Z",
-			"conversation",
-			Some(String::from("User asked Decodex to promote the accepted candidate.")),
-		);
-
-		assert!(missing_promotion_authority.is_err());
-		assert_eq!(
-			store
-				.decision_contract("decodex", candidate.contract_id())
-				.expect("contract should read")
-				.expect("contract should exist")
-				.status(),
-			DecisionContractStatus::DraftLatent
-		);
-
-		let promoted = store
-			.promote_decision_contract(
-				"decodex",
-				candidate.contract_id(),
-				DecisionPromotion::new(
-					"operator",
-					DecisionPromotionActorKind::User,
-					"2026-06-22T00:05:00Z",
-					"conversation",
-					Some(String::from("User asked Decodex to promote the accepted candidate.")),
-				)
-				.expect("promotion authority should validate"),
-			)
-			.expect("valid promotion should use existing Decision Contract semantics");
-
-		assert_eq!(promoted.status(), DecisionContractStatus::AcceptedPromoted);
-
-		let reaccept_after_promote = store
-			.accept_autonomy_proposal_as_decision_contract_candidate(
-				"decodex",
-				&proposal_id,
-				bridge_authority(),
-			)
-			.expect_err("accepted promoted contract must not be overwritten by proposal re-accept");
-
-		assert!(reaccept_after_promote.to_string().contains("will not replace"));
-		assert_eq!(
-			store
-				.decision_contract("decodex", candidate.contract_id())
-				.expect("contract should read")
-				.expect("contract should exist")
-				.status(),
-			DecisionContractStatus::AcceptedPromoted
-		);
-	}
-
-	#[test]
-	fn autonomy_decision_bridge_reaccept_refuses_generated_link_replacement() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let objective = store_accepted_objective(&store);
-		let signal = store
-			.record_autonomy_signal("decodex", runtime_signal())
-			.expect("signal should store")
-			.signal()
-			.clone();
-		let proposal =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[signal], compile_input())
-				.expect("proposal should compile");
-		let proposal_id = proposal.id().to_owned();
-
-		store.record_autonomy_proposal("decodex", proposal).expect("proposal should persist");
-
-		let candidate = store
-			.accept_autonomy_proposal_as_decision_contract_candidate(
-				"decodex",
-				&proposal_id,
-				bridge_authority(),
-			)
-			.expect("accepted proposal should become a latent Decision Contract");
-		let mut linked_contract = candidate.contract().clone();
-
-		linked_contract
-			.link_generated_execution_surfaces(["id-XY-G1"], ["XY-G1"], ["node-1"])
-			.expect("test contract links should validate");
-		store
-			.upsert_decision_contract(
-				"decodex",
-				candidate.source_issue_id(),
-				linked_contract.clone(),
-			)
-			.expect("linked contract should persist");
-
-		let reaccept_after_links = store
-			.accept_autonomy_proposal_as_decision_contract_candidate(
-				"decodex",
-				&proposal_id,
-				bridge_authority(),
-			)
-			.expect_err("generated execution links must not be overwritten");
-
-		assert!(reaccept_after_links.to_string().contains("will not replace"));
-
-		let readback = store
-			.decision_contract("decodex", candidate.contract_id())
-			.expect("contract should read")
-			.expect("contract should exist");
-
-		assert_eq!(readback.contract().links().generated_issue_identifiers(), &["XY-G1"]);
-		assert_eq!(readback.contract(), &linked_contract);
-	}
-
-	#[test]
-	fn autonomy_decision_bridge_rejected_and_needs_human_proposals_remain_non_executable() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let objective = store_accepted_objective(&store);
-		let signal = store
-			.record_autonomy_signal("decodex", runtime_signal())
-			.expect("signal should store")
-			.signal()
-			.clone();
-		let mut rejected_input = compile_input();
-
-		rejected_input.intended_surface = String::from("scripts/unowned.rs");
-
-		let rejected = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			slice::from_ref(&signal),
-			rejected_input,
-		)
-		.expect("rejected proposal should compile");
-		let rejected_id = rejected.id().to_owned();
-
-		assert_eq!(rejected.state(), AutonomyProposalState::Rejected);
-
-		store
-			.record_autonomy_proposal("decodex", rejected)
-			.expect("rejected proposal should persist");
-
-		assert!(
-			store
-				.accept_autonomy_proposal_as_decision_contract_candidate(
-					"decodex",
-					&rejected_id,
-					bridge_authority(),
-				)
-				.is_err()
-		);
-
-		let mut contradiction_input = signal_input();
-
-		contradiction_input.contradictions =
-			vec![String::from("Runtime and tracker authority disagree.")];
-
-		let contradiction_signal = store
-			.record_autonomy_signal(
-				"decodex",
-				AutonomySignal::runtime_health(contradiction_input)
-					.expect("signal should validate"),
-			)
-			.expect("contradiction signal should store")
-			.signal()
-			.clone();
-		let needs_human = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			&[contradiction_signal],
-			compile_input(),
-		)
-		.expect("needs-human proposal should compile");
-		let needs_human_id = needs_human.id().to_owned();
-
-		assert_eq!(needs_human.state(), AutonomyProposalState::NeedsHumanDecision);
-
-		store
-			.record_autonomy_proposal("decodex", needs_human)
-			.expect("needs-human proposal should persist");
-
-		assert!(
-			store
-				.accept_autonomy_proposal_as_decision_contract_candidate(
-					"decodex",
-					&needs_human_id,
-					bridge_authority(),
-				)
-				.is_err()
-		);
-		assert!(
-			store
-				.list_decision_contracts_for_project("decodex")
-				.expect("contracts should list")
-				.is_empty()
-		);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-		assert!(store.list_program_intake_plans("decodex").expect("intake plans").is_empty());
-	}
-
-	fn assert_external_agent_policy_authority_validation() {
-		let self_accept_without_policy = AutonomyProposalDecisionBridgeAuthority::new(
-			"subagent",
-			AutonomyProposalAuthorityActorKind::User,
-			"2026-06-22T00:03:00Z",
-			"agent-output",
-			"Agent accepted its own proposal.",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			None,
-		);
-
-		assert!(self_accept_without_policy.is_err());
-
-		let wrong_actor_policy = AutonomyProposalDecisionBridgeAuthority::new(
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			"2026-06-22T00:03:00Z",
-			"runtime-policy",
-			"Agent tried to rely on another actor's policy.",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			Some(accepted_project_policy(
-				"other-agent",
-				AutonomyProposalAuthorityActorKind::ExternalAgent,
-				"runtime-policy",
-			)),
-		);
-
-		assert!(wrong_actor_policy.is_err());
-
-		let wrong_source_policy = AutonomyProposalDecisionBridgeAuthority::new(
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			"2026-06-22T00:03:00Z",
-			"runtime-policy",
-			"Agent tried to rely on a policy for a different source.",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			Some(accepted_project_policy(
-				"subagent",
-				AutonomyProposalAuthorityActorKind::ExternalAgent,
-				"manual-only",
-			)),
-		);
-
-		assert!(wrong_source_policy.is_err());
-
-		let missing_acceptance_scope = AutonomyProposalAcceptedProjectPolicy::new(
-			"decodex",
-			"quality-autonomy",
-			1,
-			"quality-autonomy-policy",
-			"1",
-			"decodex.runtime_policy:quality-autonomy-policy@1",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			vec![String::from("runtime-policy")],
-			vec![String::from("other_scope")],
-		);
-
-		assert!(missing_acceptance_scope.is_err());
-	}
-
-	fn assert_policy_objective_lineage_required(store: &StateStore, proposal_id: &str) {
-		let wrong_objective_policy = AutonomyProposalAcceptedProjectPolicy::new(
-			"decodex",
-			"other-objective",
-			1,
-			"quality-autonomy-policy",
-			"1",
-			"decodex.runtime_policy:quality-autonomy-policy@1",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			vec![String::from("runtime-policy")],
-			vec![String::from(super::AUTONOMY_PROPOSAL_ACCEPTANCE_SCOPE)],
-		)
-		.expect("wrong objective policy shape should still validate");
-		let wrong_objective_authority = AutonomyProposalDecisionBridgeAuthority::new(
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			"2026-06-22T00:03:00Z",
-			"runtime-policy",
-			"Accepted project policy references the wrong objective.",
-			"subagent",
-			AutonomyProposalAuthorityActorKind::ExternalAgent,
-			Some(wrong_objective_policy),
-		)
-		.expect("authority validates before proposal lineage is checked");
-		let wrong_objective_accept = store
-			.accept_autonomy_proposal_as_decision_contract_candidate(
-				"decodex",
-				proposal_id,
-				wrong_objective_authority,
-			)
-			.expect_err("policy must match proposal objective lineage");
-
-		assert!(wrong_objective_accept.to_string().contains("does not match proposal"));
-	}
-
-	#[test]
-	fn autonomy_decision_bridge_external_agent_self_accept_requires_project_policy() {
-		assert_external_agent_policy_authority_validation();
-
-		let store = StateStore::open_in_memory().expect("store should open");
-		let objective = store_accepted_objective(&store);
-		let mut input = signal_input();
-
-		input.source_type = AutonomySignalSourceType::Agent;
-
-		let signal = store
-			.record_autonomy_signal(
-				"decodex",
-				AutonomySignal::runtime_health(input).expect("agent signal should validate"),
-			)
-			.expect("agent signal should store")
-			.signal()
-			.clone();
-		let proposal =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[signal], compile_input())
-				.expect("proposal should compile");
-		let proposal_id = proposal.id().to_owned();
-
-		store.record_autonomy_proposal("decodex", proposal).expect("proposal should persist");
-
-		assert_policy_objective_lineage_required(&store, &proposal_id);
-
-		let candidate = store
-			.accept_autonomy_proposal_as_decision_contract_candidate(
-				"decodex",
-				&proposal_id,
-				runtime_policy_bridge_authority(),
-			)
-			.expect("policy-backed external acceptance should bridge to latent contract");
-
-		assert_eq!(candidate.status(), DecisionContractStatus::DraftLatent);
-		assert!(candidate.contract().promotion().is_none());
-	}
-
-	#[test]
-	fn autonomy_proposal_id_ignores_timestamps_signal_order_warning_order_and_challenges() {
-		let objective = objective_fixture();
-		let signal = runtime_signal();
-		let mut second_input = signal_input();
-
-		second_input.source_refs = vec![String::from("status:runtime-health:secondary")];
-		second_input.evidence = vec![String::from("secondary readback")];
-
-		let second_signal =
-			AutonomySignal::runtime_health(second_input).expect("second signal should validate");
-		let mut input_a = compile_input();
-		let mut input_b = compile_input();
-
-		input_a.affected_identifiers = vec![String::from("b"), String::from("a")];
-		input_a.created_at = String::from("2026-06-22T00:01:00Z");
-		input_b.affected_identifiers = vec![String::from("a"), String::from("b")];
-		input_b.created_at = String::from("2026-06-22T00:55:00Z");
-
-		let proposal_a = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			&[signal.clone(), second_signal.clone()],
-			input_a,
-		)
-		.expect("proposal a should compile");
-		let mut proposal_b = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			&[second_signal, signal.clone(), signal],
-			input_b,
-		)
-		.expect("proposal b should compile");
-		let original_id = proposal_b.id().to_owned();
-
-		proposal_b
-			.record_challenge(AutonomyProposalChallengeInput {
-				source: AutonomyProposalChallengeSource::InlineSkeptic,
-				actor: String::from("inline"),
-				summary: String::from("Skeptic noted a possible operator wording gap."),
-				objections: Vec::new(),
-				evidence_refs: vec![String::from("challenge:inline")],
-				recorded_at: String::from("2026-06-22T00:56:00Z"),
-			})
-			.expect("challenge should record");
-
-		assert_eq!(proposal_a.id(), original_id);
-		assert_eq!(proposal_a.fingerprint(), proposal_b.fingerprint());
-		assert_eq!(proposal_b.id(), original_id);
-	}
-
-	#[test]
-	fn autonomy_proposal_refusal_reasons_are_specific_and_inspectable() {
-		let objective = objective_fixture();
-		let signal = runtime_signal();
-		let missing =
-			AutonomyProposal::compile_dry_run(None, slice::from_ref(&signal), compile_input())
-				.expect("missing objective proposal should compile as refusal");
-
-		assert_eq!(missing.state(), AutonomyProposalState::NeedsEvidence);
-		assert!(missing.has_refusal_reason(AutonomyProposalRefusalReason::MissingObjective));
-
-		let mut stale_input = signal_input();
-
-		stale_input.freshness = AutonomySignalFreshness::Stale;
-
-		let stale_signal =
-			AutonomySignal::runtime_health(stale_input).expect("stale signal should validate");
-		let stale =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[stale_signal], compile_input())
-				.expect("stale evidence proposal should compile as refusal");
-
-		assert_eq!(stale.state(), AutonomyProposalState::NeedsEvidence);
-		assert!(stale.has_refusal_reason(AutonomyProposalRefusalReason::StaleEvidence));
-
-		let mut contradiction_input = signal_input();
-
-		contradiction_input.contradictions =
-			vec![String::from("Tracker says closed while runtime says active.")];
-
-		let contradictory_signal = AutonomySignal::runtime_health(contradiction_input)
-			.expect("contradictory signal should validate");
-		let contradictory = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			&[contradictory_signal],
-			compile_input(),
-		)
-		.expect("contradictory proposal should compile as refusal");
-
-		assert_eq!(contradictory.state(), AutonomyProposalState::NeedsHumanDecision);
-		assert!(
-			contradictory
-				.has_refusal_reason(AutonomyProposalRefusalReason::UnresolvedContradiction)
-		);
-
-		let mut weakened_input = compile_input();
-
-		weakened_input.weakened_validation_or_review =
-			vec![String::from("Review evidence is older than the current head.")];
-
-		let weakened = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			slice::from_ref(&signal),
-			weakened_input,
-		)
-		.expect("weakened validation proposal should compile as refusal");
-
-		assert_eq!(weakened.state(), AutonomyProposalState::NeedsEvidence);
-		assert!(
-			weakened.has_refusal_reason(AutonomyProposalRefusalReason::WeakenedValidationReview)
-		);
-
-		let mut disallowed_surface_input = compile_input();
-
-		disallowed_surface_input.intended_surface = String::from("scripts/unowned.rs");
-
-		let disallowed_surface = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			slice::from_ref(&signal),
-			disallowed_surface_input,
-		)
-		.expect("disallowed surface proposal should compile as refusal");
-
-		assert_eq!(disallowed_surface.state(), AutonomyProposalState::Rejected);
-		assert!(
-			disallowed_surface.has_refusal_reason(AutonomyProposalRefusalReason::DisallowedSurface)
-		);
-
-		let mut traversal_surface_input = compile_input();
-
-		traversal_surface_input.intended_surface =
-			String::from("apps/decodex/src/../../scripts/unowned.rs");
-
-		let traversal_surface = AutonomyProposal::compile_dry_run(
-			Some(&objective),
-			slice::from_ref(&signal),
-			traversal_surface_input,
-		)
-		.expect("traversal surface proposal should compile as refusal");
-
-		assert_eq!(traversal_surface.state(), AutonomyProposalState::Rejected);
-		assert!(
-			traversal_surface.has_refusal_reason(AutonomyProposalRefusalReason::DisallowedSurface)
-		);
-
-		let docs_signal =
-			AutonomySignal::docs_skill_drift(signal_input()).expect("docs signal should validate");
-		let disallowed_kind =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[docs_signal], compile_input())
-				.expect("disallowed signal proposal should compile as refusal");
-
-		assert_eq!(disallowed_kind.state(), AutonomyProposalState::Rejected);
-		assert!(
-			disallowed_kind.has_refusal_reason(AutonomyProposalRefusalReason::DisallowedSignalKind)
-		);
-	}
-
-	#[test]
-	fn autonomy_proposal_rejects_promoted_state_without_decision_contract_provenance() {
-		let objective = objective_fixture();
-		let signal = runtime_signal();
-		let mut proposal =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[signal], compile_input())
-				.expect("proposal should compile");
-
-		proposal.state = AutonomyProposalState::AcceptedPromoted;
-
-		assert!(
-			proposal
-				.validate()
-				.expect_err("accepted_promoted should require promotion provenance")
-				.to_string()
-				.contains("cannot claim accepted_promoted")
-		);
-	}
-
-	#[test]
-	fn autonomy_proposal_challenge_records_objections_without_acceptance_authority() {
-		let objective = objective_fixture();
-		let signal = runtime_signal();
-		let mut proposal =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[signal], compile_input())
-				.expect("proposal should compile");
-		let proposal_id = proposal.id().to_owned();
-
-		proposal
-			.record_challenge(AutonomyProposalChallengeInput {
-				source: AutonomyProposalChallengeSource::Subagent,
-				actor: String::from("subagent"),
-				summary: String::from("Subagent challenged the evidence sufficiency."),
-				objections: vec![String::from("Needs a fresher operator status readback.")],
-				evidence_refs: vec![String::from("challenge:subagent")],
-				recorded_at: String::from("2026-06-22T00:02:00Z"),
-			})
-			.expect("challenge should record");
-
-		assert_eq!(proposal.id(), proposal_id);
-		assert_eq!(proposal.state(), AutonomyProposalState::DecisionCandidate);
-		assert_eq!(proposal.challenge_evidence().len(), 1);
-		assert!(!proposal.challenge_evidence()[0].acceptance_authority);
-		assert_eq!(
-			proposal.challenge_evidence()[0].objections,
-			["Needs a fresher operator status readback."]
-		);
-
-		let dry_run_json = serde_json::to_value(&proposal).expect("proposal should encode");
-
-		assert_eq!(dry_run_json["challenge_evidence"][0]["acceptance_authority"], false);
-		assert_eq!(
-			dry_run_json["challenge_evidence"][0]["objections"][0],
-			"Needs a fresher operator status readback."
-		);
-
-		let candidate = proposal
-			.to_decision_contract_candidate(bridge_authority())
-			.expect("challenge objections should remain promotion constraints");
-
-		assert!(candidate.accepted_authority().constraints().contains(&String::from(
-			"Challenge promotion constraint: Needs a fresher operator status readback."
-		)));
-		assert!(
-			candidate
-				.accepted_authority()
-				.objections()
-				.contains(&String::from("Needs a fresher operator status readback."))
-		);
-	}
-
-	#[test]
-	fn autonomy_proposal_store_round_trips_without_execution_authority_side_effects() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let objective = store_accepted_objective(&store);
-
-		objective.supersession().expect_none("accepted fixture must not have supersession");
-
-		let signal = store
-			.record_autonomy_signal("decodex", runtime_signal())
-			.expect("signal should store")
-			.signal()
-			.clone();
-		let proposal =
-			AutonomyProposal::compile_dry_run(Some(&objective), &[signal], compile_input())
-				.expect("proposal should compile");
-		let stored = store
-			.record_autonomy_proposal("decodex", proposal.clone())
-			.expect("proposal should persist");
-
-		assert_eq!(stored.proposal(), &proposal);
-		assert_eq!(
-			store
-				.autonomy_proposal("decodex", proposal.id())
-				.expect("proposal read should work")
-				.expect("proposal should exist")
-				.proposal(),
-			&proposal
-		);
-		assert!(
-			store
-				.list_decision_contracts_for_project("decodex")
-				.expect("decision contracts should list")
-				.is_empty()
-		);
-		assert!(store.list_execution_programs("decodex").expect("programs should list").is_empty());
-		assert!(
-			store
-				.list_program_intake_plans("decodex")
-				.expect("intake plans should list")
-				.is_empty()
-		);
-	}
-
-	#[test]
-	fn autonomy_proposal_sqlite_round_trips_full_dry_run_record() {
-		let tempdir = tempfile::tempdir().expect("tempdir should create");
-		let db_path = tempdir.path().join("runtime.sqlite3");
-		let stored_proposal = {
-			let store = StateStore::open(&db_path).expect("store should open");
-			let objective = store_accepted_objective(&store);
-			let signal = store
-				.record_autonomy_signal("decodex", runtime_signal())
-				.expect("signal should store")
-				.signal()
-				.clone();
-			let mut proposal =
-				AutonomyProposal::compile_dry_run(Some(&objective), &[signal], compile_input())
-					.expect("proposal should compile");
-
-			proposal
-				.record_challenge(AutonomyProposalChallengeInput {
-					source: AutonomyProposalChallengeSource::Subagent,
-					actor: String::from("subagent"),
-					summary: String::from("Subagent challenged the evidence sufficiency."),
-					objections: vec![String::from("Needs a fresher operator status readback.")],
-					evidence_refs: vec![String::from("challenge:subagent")],
-					recorded_at: String::from("2026-06-22T00:02:00Z"),
-				})
-				.expect("challenge should record");
-			store
-				.record_autonomy_proposal("decodex", proposal.clone())
-				.expect("proposal should persist");
-
-			proposal
-		};
-		let reopened = StateStore::open(&db_path).expect("store should reopen");
-		let readback = reopened
-			.autonomy_proposal("decodex", stored_proposal.id())
-			.expect("proposal read should work")
-			.expect("proposal should exist");
-
-		assert_eq!(readback.proposal(), &stored_proposal);
-		assert_eq!(readback.state(), AutonomyProposalState::DecisionCandidate);
-		assert_eq!(
-			reopened
-				.recent_autonomy_proposals_for_project("decodex", 1)
-				.expect("recent proposals should list")[0]
-				.proposal(),
-			&stored_proposal
-		);
-		assert!(
-			reopened
-				.list_decision_contracts_for_project("decodex")
-				.expect("decision contracts should list")
-				.is_empty()
-		);
-		assert!(
-			reopened.list_execution_programs("decodex").expect("programs should list").is_empty()
-		);
-		assert!(
-			reopened
-				.list_program_intake_plans("decodex")
-				.expect("intake plans should list")
-				.is_empty()
-		);
 	}
 }
