+++ apps/decodex/src/program_intake.rs
@@ -30,12 +30,17 @@ use crate::{
 	workflow::WorkflowDocument,
 };
 
+mod goal;
+mod issue_batch;
 mod render;
+#[cfg(test)] mod tests;
 
 use self::render::{
 	generated_issue_private_identifiers, render_goal_issue_brief, validate_generated_issue_text,
 };
 pub(crate) use self::render::{render_goal_intake_report, render_issue_batch_intake_report};
+#[allow(clippy::wildcard_imports)] use goal::*;
+#[allow(clippy::wildcard_imports)] use issue_batch::*;
 
 /// CLI/runtime request for issue-batch Program Intake.
 pub(crate) struct IssueBatchIntakeCommandRequest<'a> {
@@ -551,1950 +556,4 @@ where
 		counts,
 		issues: rows,
 	})
-}
-
-fn ensure_goal_intake_authority(contract: &DecisionContract) -> Result<()> {
-	if contract.status() != DecisionContractStatus::AcceptedPromoted {
-		eyre::bail!(
-			"Decision Contract `{}` is `{}`; goal intake requires accepted execution authority.",
-			contract.contract_id(),
-			contract.status().as_str()
-		);
-	}
-	if !contract.execution_readiness().ready_for_issue_shaping() {
-		eyre::bail!(
-			"Decision Contract `{}` is not ready for issue shaping.",
-			contract.contract_id()
-		);
-	}
-	if !contract.execution_readiness().missing_decisions().is_empty() {
-		eyre::bail!(
-			"Decision Contract `{}` still has unresolved decisions.",
-			contract.contract_id()
-		);
-	}
-	if contract.execution_readiness().proposed_issues().is_empty() {
-		eyre::bail!(
-			"Decision Contract `{}` has no structured proposed issues to materialize.",
-			contract.contract_id()
-		);
-	}
-
-	Ok(())
-}
-
-fn goal_issue_plans(contract: &DecisionContract, program_id: &str) -> Result<Vec<GoalIssuePlan>> {
-	let mut plans = Vec::new();
-
-	for (index, issue) in contract.execution_readiness().proposed_issues().iter().enumerate() {
-		let node_id = goal_node_id(contract.contract_id(), index, issue.key());
-		let title = issue.title().to_owned();
-		let objective = issue.objective().to_owned();
-		let acceptance = issue.acceptance().to_vec();
-		let validation = issue.validation().to_vec();
-		let risk = issue.risk().to_vec();
-		let dependencies = issue.dependencies().to_vec();
-		let conflict_domains = goal_proposed_issue_conflict_domains(issue)?;
-		let description = render_goal_issue_brief(GoalIssueBriefInput {
-			contract,
-			objective: &objective,
-			dependencies: &dependencies,
-			conflict_domains: &conflict_domains,
-			acceptance: &acceptance,
-			validation: &validation,
-			risk: &risk,
-		})?;
-		let private_identifiers =
-			generated_issue_private_identifiers(contract, program_id, &node_id);
-		let private_identifier_refs =
-			private_identifiers.iter().map(String::as_str).collect::<Vec<_>>();
-
-		validate_generated_issue_text(&title, &description, &private_identifier_refs)?;
-
-		plans.push(GoalIssuePlan {
-			key: issue.key().to_owned(),
-			node_id,
-			title,
-			objective,
-			stage: parse_goal_stage(issue.stage())?,
-			queue_intent: parse_goal_queue_intent(issue.queue_intent())?,
-			description,
-			dependencies,
-			dependency_node_ids: Vec::new(),
-			conflict_domains,
-			acceptance,
-			validation,
-			risk,
-		});
-	}
-
-	bind_goal_dependency_node_ids(&mut plans);
-
-	Ok(plans)
-}
-
-fn bind_goal_dependency_node_ids(plans: &mut [GoalIssuePlan]) {
-	let node_ids_by_key = plans
-		.iter()
-		.map(|plan| (plan.key.clone(), plan.node_id.clone()))
-		.collect::<BTreeMap<_, _>>();
-
-	for plan in plans {
-		plan.dependency_node_ids = plan
-			.dependencies
-			.iter()
-			.map(|dependency| {
-				node_ids_by_key.get(dependency).cloned().unwrap_or_else(|| dependency.to_owned())
-			})
-			.collect();
-	}
-}
-
-fn linked_goal_issues<T>(
-	tracker: &T,
-	contract: &DecisionContract,
-	plan_count: usize,
-) -> Result<Vec<Option<TrackerIssue>>>
-where
-	T: IssueTracker + ?Sized,
-{
-	let mut linked = Vec::with_capacity(plan_count);
-
-	for index in 0..plan_count {
-		let issue = match contract.links().generated_issue_identifiers().get(index) {
-			Some(identifier) =>
-				Some(tracker.get_issue_by_identifier(identifier)?.ok_or_else(|| {
-					eyre::eyre!(
-						"Generated issue link `{identifier}` for Decision Contract `{}` did not resolve.",
-						contract.contract_id()
-					)
-				})?),
-			None => None,
-		};
-
-		linked.push(issue);
-	}
-
-	Ok(linked)
-}
-
-fn goal_intake_anchor<T>(
-	tracker: &T,
-	workflow: &WorkflowDocument,
-	team_issue_identifier: Option<String>,
-) -> Result<GoalIntakeAnchor>
-where
-	T: IssueTracker + ?Sized,
-{
-	let identifier = team_issue_identifier.ok_or_else(|| {
-		eyre::eyre!(
-			"Goal intake apply requires a source issue on the Decision Contract or --team-issue <ISSUE>."
-		)
-	})?;
-	let issue = tracker
-		.get_issue_by_identifier(&identifier)?
-		.ok_or_else(|| eyre::eyre!("Team anchor issue `{identifier}` did not resolve."))?;
-	let (state_id, _state_name) = workflow
-		.frontmatter()
-		.tracker()
-		.startable_states()
-		.iter()
-		.find_map(|state_name| {
-			issue
-				.state_id_for_name(state_name)
-				.map(|state_id| (state_id.to_owned(), state_name.as_str()))
-		})
-		.ok_or_else(|| {
-			eyre::eyre!(
-				"Team anchor issue `{}` does not expose any configured startable state.",
-				issue.identifier
-			)
-		})?;
-
-	Ok(GoalIntakeAnchor { team_id: issue.team.id, state_id })
-}
-
-fn apply_goal_issues_and_link_contract<T>(
-	input: ApplyGoalIssuesInput<'_, T>,
-) -> Result<(Vec<TrackerIssue>, DecisionContract)>
-where
-	T: IssueTracker + ?Sized,
-{
-	let ApplyGoalIssuesInput {
-		state_store,
-		service_id,
-		source_issue_id,
-		tracker,
-		contract,
-		plans,
-		linked_issues,
-		anchor,
-	} = input;
-	let mut issues = Vec::with_capacity(plans.len());
-	let mut linked_contract = contract.clone();
-
-	for (plan, linked_issue) in plans.iter().zip(linked_issues) {
-		let issue = match linked_issue {
-			Some(issue) => tracker.update_issue_brief(
-				&issue.id,
-				&TrackerIssueBriefUpdate {
-					title: plan.title.clone(),
-					description: plan.description.clone(),
-				},
-			)?,
-			None => tracker.create_issue(&TrackerIssueCreate {
-				team_id: anchor.team_id.clone(),
-				title: plan.title.clone(),
-				description: plan.description.clone(),
-				state_id: Some(anchor.state_id.clone()),
-			})?,
-		};
-
-		issues.push(issue);
-
-		linked_contract =
-			linked_goal_contract_for_apply_progress(contract, plans, linked_issues, &issues)?;
-
-		state_store.upsert_decision_contract(
-			service_id,
-			source_issue_id,
-			linked_contract.clone(),
-		)?;
-	}
-
-	Ok((issues, linked_contract))
-}
-
-fn linked_goal_contract_for_apply_progress(
-	contract: &DecisionContract,
-	plans: &[GoalIssuePlan],
-	linked_issues: &[Option<TrackerIssue>],
-	applied_issues: &[TrackerIssue],
-) -> Result<DecisionContract> {
-	let mut linked_contract = contract.clone();
-	let mut issue_ids = Vec::new();
-	let mut issue_identifiers = Vec::new();
-	let mut node_ids = Vec::new();
-
-	for (index, plan) in plans.iter().enumerate() {
-		let issue =
-			applied_issues.get(index).or_else(|| linked_issues.get(index).and_then(Option::as_ref));
-
-		if let Some(issue) = issue {
-			issue_ids.push(issue.id.clone());
-			issue_identifiers.push(issue.identifier.clone());
-
-			let node_id = if applied_issues.get(index).is_some() {
-				plan.node_id.clone()
-			} else {
-				contract
-					.links()
-					.execution_program_node_ids()
-					.get(index)
-					.cloned()
-					.unwrap_or_else(|| plan.node_id.clone())
-			};
-
-			node_ids.push(node_id);
-		}
-	}
-
-	linked_contract.link_generated_execution_surfaces(issue_ids, issue_identifiers, node_ids)?;
-
-	Ok(linked_contract)
-}
-
-fn goal_execution_program(
-	service_id: &str,
-	program_id: &str,
-	contract: &DecisionContract,
-	plans: &[GoalIssuePlan],
-	issues: &[TrackerIssue],
-	workflow: &WorkflowDocument,
-) -> Result<ExecutionProgram> {
-	let nodes = plans
-		.iter()
-		.zip(issues)
-		.map(|(plan, issue)| goal_program_node(service_id, contract, plan, issue, workflow))
-		.collect::<Result<Vec<_>>>()?;
-
-	ExecutionProgram::from_accepted_contract(program_id, service_id, contract, nodes)
-}
-
-fn goal_program_node(
-	service_id: &str,
-	contract: &DecisionContract,
-	plan: &GoalIssuePlan,
-	issue: &TrackerIssue,
-	workflow: &WorkflowDocument,
-) -> Result<ExecutionProgramNode> {
-	let dependencies = plan
-		.dependency_node_ids
-		.iter()
-		.map(ExecutionProgramDependency::new)
-		.collect::<Result<Vec<_>>>()?;
-	let mapping = goal_issue_mapping(service_id, issue, workflow)?;
-
-	ExecutionProgramNode::new(
-		plan.node_id.clone(),
-		plan.stage,
-		plan.objective.clone(),
-		plan.queue_intent,
-	)?
-	.with_objective_lineage(goal_objective_lineage(contract))?
-	.with_dependencies(dependencies)?
-	.with_conflict_domains(plan.conflict_domains.clone())?
-	.with_acceptance_expectations(plan.acceptance.clone())?
-	.with_validation_expectations(plan.validation.clone())?
-	.with_linear_issue(mapping)
-}
-
-fn goal_issue_mapping(
-	service_id: &str,
-	issue: &TrackerIssue,
-	workflow: &WorkflowDocument,
-) -> Result<ExecutionLinearIssueMapping> {
-	let active_label = tracker::automation_active_label(service_id);
-	let tracker_policy = workflow.frontmatter().tracker();
-
-	Ok(ExecutionLinearIssueMapping::new(&issue.id, &issue.identifier, &issue.state.name)?
-		.with_active_label(issue.has_label(&active_label))
-		.with_opt_out_label(issue.has_label(tracker_policy.opt_out_label()))
-		.with_needs_attention_label(issue.has_label(tracker_policy.needs_attention_label()))
-		.with_generic_dispatch_briefing(issue_has_generic_dispatch_briefing(issue)))
-}
-
-fn applied_goal_issue_rows(
-	plans: &[GoalIssuePlan],
-	issues: &[TrackerIssue],
-	linked_issues: &[Option<TrackerIssue>],
-	evaluation: &ExecutionProgramEvaluation,
-) -> Vec<GoalIntakeIssueReport> {
-	plans
-		.iter()
-		.zip(issues)
-		.zip(linked_issues)
-		.map(|((plan, issue), linked)| {
-			let evaluation = evaluation.nodes().iter().find(|node| node.node_id() == plan.node_id);
-
-			goal_issue_report_row(
-				plan,
-				Some(issue),
-				if linked.is_some() {
-					GoalIntakeIssueAction::Updated
-				} else {
-					GoalIntakeIssueAction::Created
-				},
-				evaluation.and_then(ExecutionNodeEvaluation::dispatch_action),
-				evaluation.map_or_else(Vec::new, |node| node.reasons().to_vec()),
-			)
-		})
-		.collect()
-}
-
-fn dry_run_goal_issue_rows(
-	plans: &[GoalIssuePlan],
-	linked_issues: &[Option<TrackerIssue>],
-) -> Vec<GoalIntakeIssueReport> {
-	plans
-		.iter()
-		.zip(linked_issues)
-		.map(|(plan, linked)| {
-			let action = if linked.is_some() {
-				GoalIntakeIssueAction::WouldUpdate
-			} else {
-				GoalIntakeIssueAction::WouldCreate
-			};
-			let reason = match action {
-				GoalIntakeIssueAction::WouldCreate =>
-					"apply will create a normal Linear issue and persist a mapped program node",
-				GoalIntakeIssueAction::WouldUpdate =>
-					"apply will update the linked normal Linear issue and persist a mapped program node",
-				GoalIntakeIssueAction::Created | GoalIntakeIssueAction::Updated =>
-					"apply already materialized this issue",
-			};
-
-			goal_issue_report_row(plan, linked.as_ref(), action, None, vec![reason.to_owned()])
-		})
-		.collect()
-}
-
-fn goal_issue_report_row(
-	plan: &GoalIssuePlan,
-	issue: Option<&TrackerIssue>,
-	action: GoalIntakeIssueAction,
-	dispatch_action: Option<ExecutionDispatchAction>,
-	reasons: Vec<String>,
-) -> GoalIntakeIssueReport {
-	GoalIntakeIssueReport {
-		node_id: plan.node_id.clone(),
-		title: plan.title.clone(),
-		objective: plan.objective.clone(),
-		issue_id: issue.map(|issue| issue.id.clone()),
-		issue_identifier: issue.map(|issue| issue.identifier.clone()),
-		action,
-		queue_intent: plan.queue_intent.as_str().to_owned(),
-		dispatch_action: dispatch_action.map(dispatch_action_name),
-		dependencies: plan.dependencies.clone(),
-		conflict_domains: conflict_domain_labels(&plan.conflict_domains),
-		acceptance: plan.acceptance.clone(),
-		validation: plan.validation.clone(),
-		reasons,
-	}
-}
-
-fn goal_objective_lineage(contract: &DecisionContract) -> Vec<String> {
-	let mut lineage = vec![
-		format!("Accepted Decision Contract `{}`.", contract.contract_id()),
-		format!("Source intent: {}", contract.source_intent().summary()),
-	];
-
-	lineage.extend(contract.accepted_authority().accepted_objectives().iter().cloned());
-
-	lineage
-}
-
-fn goal_proposed_issue_conflict_domains(
-	issue: &DecisionProposedIssue,
-) -> Result<Vec<ExecutionConflictDomain>> {
-	issue.conflict_domains().iter().map(|domain| parse_goal_conflict_domain(domain)).collect()
-}
-
-fn parse_goal_conflict_domain(domain: &str) -> Result<ExecutionConflictDomain> {
-	let domain = domain.trim();
-	let (kind, key) = domain.split_once(':').unwrap_or(("module", domain));
-	let kind = match kind {
-		"file" => ExecutionConflictDomainKind::File,
-		"module" => ExecutionConflictDomainKind::Module,
-		"state" => ExecutionConflictDomainKind::State,
-		"credentials" => ExecutionConflictDomainKind::Credentials,
-		"tracker_ownership" => ExecutionConflictDomainKind::TrackerOwnership,
-		"review_surface" => ExecutionConflictDomainKind::ReviewSurface,
-		_ => return ExecutionConflictDomain::new(ExecutionConflictDomainKind::Module, domain),
-	};
-
-	ExecutionConflictDomain::new(kind, key)
-}
-
-fn parse_goal_stage(stage: &str) -> Result<ExecutionProgramNodeStage> {
-	match stage {
-		"research" => Ok(ExecutionProgramNodeStage::Research),
-		"design" => Ok(ExecutionProgramNodeStage::Design),
-		"spec" => Ok(ExecutionProgramNodeStage::Spec),
-		"schema" => Ok(ExecutionProgramNodeStage::Schema),
-		"runtime" => Ok(ExecutionProgramNodeStage::Runtime),
-		"plugin" => Ok(ExecutionProgramNodeStage::Plugin),
-		"eval" => Ok(ExecutionProgramNodeStage::Eval),
-		"handoff" => Ok(ExecutionProgramNodeStage::Handoff),
-		_ => eyre::bail!("Unsupported proposed issue stage `{stage}`."),
-	}
-}
-
-fn parse_goal_queue_intent(queue_intent: &str) -> Result<ExecutionQueueIntent> {
-	match queue_intent {
-		"not_ready" => Ok(ExecutionQueueIntent::NotReady),
-		"ready_to_queue" => Ok(ExecutionQueueIntent::ReadyToQueue),
-		"queued" => Ok(ExecutionQueueIntent::Queued),
-		"active" => Ok(ExecutionQueueIntent::Active),
-		"paused" => Ok(ExecutionQueueIntent::Paused),
-		"done" => Ok(ExecutionQueueIntent::Done),
-		"canceled" => Ok(ExecutionQueueIntent::Canceled),
-		_ => eyre::bail!("Unsupported proposed issue queue_intent `{queue_intent}`."),
-	}
-}
-
-fn conflict_domain_labels(domains: &[ExecutionConflictDomain]) -> Vec<String> {
-	let mut labels = domains
-		.iter()
-		.map(|domain| format!("{}:{}", domain.kind().as_str(), domain.key()))
-		.collect::<Vec<_>>();
-
-	labels.sort();
-	labels.dedup();
-
-	labels
-}
-
-fn goal_program_id(service_id: &str, contract_id: &str) -> String {
-	format!("goal-{service_id}-{}", stable_slug(contract_id, 48))
-}
-
-fn goal_node_id(contract_id: &str, index: usize, objective: &str) -> String {
-	format!("goal:{}:{:02}-{}", stable_slug(contract_id, 32), index + 1, stable_slug(objective, 32))
-}
-
-fn stable_slug(value: &str, max_len: usize) -> String {
-	let mut slug = String::new();
-	let mut previous_dash = false;
-
-	for character in value.chars() {
-		if character.is_ascii_alphanumeric() {
-			slug.push(character.to_ascii_lowercase());
-
-			previous_dash = false;
-		} else if !previous_dash && !slug.is_empty() {
-			slug.push('-');
-
-			previous_dash = true;
-		}
-		if slug.len() >= max_len {
-			break;
-		}
-	}
-
-	while slug.ends_with('-') {
-		slug.pop();
-	}
-
-	if slug.is_empty() { String::from("goal") } else { slug }
-}
-
-fn resolve_intake_project_config_path(
-	config_path: Option<&Path>,
-	project_id: Option<&str>,
-	state_store: &StateStore,
-) -> Result<PathBuf> {
-	if let Some(config_path) = config_path {
-		return ServiceConfig::resolve_project_config_path(config_path);
-	}
-	if let Some(project_id) = project_id {
-		let Some(project) = state_store
-			.list_projects()?
-			.into_iter()
-			.find(|project| project.service_id() == project_id)
-		else {
-			eyre::bail!(
-				"Decodex project `{project_id}` is not registered; pass --config <PROJECT_DIR>."
-			);
-		};
-
-		return Ok(project.config_path().to_path_buf());
-	}
-
-	runtime::registered_config_path_for_cwd(state_store, &env::current_dir()?)?.ok_or_else(|| {
-		eyre::eyre!(
-			"Current directory is not registered to a Decodex project; pass --config <PROJECT_DIR> or --project <SERVICE_ID>."
-		)
-	})
-}
-
-fn register_intake_project_config_for_persist(
-	state_store: &StateStore,
-	config_path: &Path,
-	persist: bool,
-) -> Result<()> {
-	if persist {
-		runtime::register_project_config(state_store, config_path, true)?;
-	}
-
-	Ok(())
-}
-
-fn normalize_issue_identifiers(issue_identifiers: Vec<String>) -> Result<Vec<String>> {
-	let mut normalized = issue_identifiers
-		.into_iter()
-		.map(|identifier| identifier.trim().to_owned())
-		.filter(|identifier| !identifier.is_empty())
-		.collect::<Vec<_>>();
-
-	normalized.sort();
-	normalized.dedup();
-
-	if normalized.is_empty() {
-		eyre::bail!("Issue-batch intake requires at least one issue identifier.");
-	}
-
-	Ok(normalized)
-}
-
-fn issue_batch_fingerprint(
-	service_id: &str,
-	issue_identifiers: &[String],
-	resolved: &BTreeMap<String, TrackerIssue>,
-) -> String {
-	let mut digest = Sha256::new();
-
-	digest.update(service_id.as_bytes());
-
-	for identifier in issue_identifiers {
-		digest.update(b"\0identifier:");
-		digest.update(identifier.as_bytes());
-
-		if let Some(issue) = resolved.get(identifier) {
-			digest.update(b"\0issue:");
-			digest.update(issue.id.as_bytes());
-			digest.update(b"\0state:");
-			digest.update(issue.state.name.as_bytes());
-			digest.update(b"\0updated:");
-			digest.update(issue.updated_at.as_bytes());
-		}
-	}
-
-	digest.finalize().iter().map(|byte| format!("{byte:02x}")).collect()
-}
-
-fn issue_batch_program_id(service_id: &str, fingerprint: &str) -> String {
-	format!("issue-batch-{service_id}-{}", &fingerprint[..16])
-}
-
-fn node_id_for_issue(identifier: &str) -> String {
-	format!("issue:{identifier}")
-}
-
-fn unmapped_node(identifier: &str) -> Result<ExecutionProgramNode> {
-	ExecutionProgramNode::new(
-		format!("unmapped:{identifier}"),
-		ExecutionProgramNodeStage::Runtime,
-		format!("Resolve supplied Linear issue identifier `{identifier}` before dispatch."),
-		ExecutionQueueIntent::NotReady,
-	)?
-	.with_acceptance_expectations([format!(
-		"`{identifier}` maps to a normal Linear issue before execution."
-	)])?
-	.with_validation_expectations([String::from("Tracker lookup succeeds before queue intent.")])
-}
-
-fn issue_facts<T>(
-	tracker: &T,
-	workflow: &WorkflowDocument,
-	issue: &TrackerIssue,
-	active_label: &str,
-) -> Result<IssueFacts>
-where
-	T: IssueTracker + ?Sized,
-{
-	let tracker_policy = workflow.frontmatter().tracker();
-	let has_active_label =
-		tracker::issue_has_label_with_server_confirmation(tracker, issue, active_label)?;
-	let has_opt_out_label = tracker::issue_has_label_with_server_confirmation(
-		tracker,
-		issue,
-		tracker_policy.opt_out_label(),
-	)?;
-	let has_needs_attention_label = tracker::issue_has_label_with_server_confirmation(
-		tracker,
-		issue,
-		tracker_policy.needs_attention_label(),
-	)?;
-	let has_open_blockers =
-		issue.blockers.iter().any(|blocker| !state_name_is_terminal(&blocker.state.name, workflow));
-
-	Ok(IssueFacts {
-		has_active_label,
-		has_opt_out_label,
-		has_needs_attention_label,
-		has_generic_dispatch_briefing: issue_has_generic_dispatch_briefing(issue),
-		has_open_blockers,
-	})
-}
-
-fn issue_node(
-	issue: &TrackerIssue,
-	facts: &IssueFacts,
-	workflow: &WorkflowDocument,
-	supplied_node_ids: &BTreeMap<String, String>,
-) -> Result<ExecutionProgramNode> {
-	let queue_intent = issue_queue_intent(issue, facts, workflow);
-	let mut mapping =
-		ExecutionLinearIssueMapping::new(&issue.id, &issue.identifier, &issue.state.name)?;
-
-	mapping = mapping
-		.with_active_label(facts.has_active_label)
-		.with_opt_out_label(facts.has_opt_out_label)
-		.with_needs_attention_label(facts.has_needs_attention_label)
-		.with_open_tracker_blockers(facts.has_open_blockers)
-		.with_generic_dispatch_briefing(facts.has_generic_dispatch_briefing);
-
-	ExecutionProgramNode::new(
-		node_id_for_issue(&issue.identifier),
-		ExecutionProgramNodeStage::Runtime,
-		issue.title.clone(),
-		queue_intent,
-	)?
-	.with_objective_lineage([format!("Issue-batch intake supplied `{}`.", issue.identifier)])?
-	.with_dependencies(issue_dependencies(issue, supplied_node_ids)?)?
-	.with_conflict_domains(issue_conflict_domains(issue)?)?
-	.with_acceptance_expectations([format!(
-		"`{}` remains a normal Linear issue with an executable brief.",
-		issue.identifier
-	)])?
-	.with_validation_expectations([String::from(
-		"Run the issue-specific repository validation before review handoff.",
-	)])?
-	.with_linear_issue(mapping)
-}
-
-fn issue_queue_intent(
-	issue: &TrackerIssue,
-	facts: &IssueFacts,
-	workflow: &WorkflowDocument,
-) -> ExecutionQueueIntent {
-	if state_name_is_terminal(&issue.state.name, workflow) {
-		return ExecutionQueueIntent::Done;
-	}
-	if facts.has_active_label {
-		return ExecutionQueueIntent::Active;
-	}
-	if facts.has_opt_out_label {
-		return ExecutionQueueIntent::NotReady;
-	}
-	if !workflow
-		.frontmatter()
-		.tracker()
-		.startable_states()
-		.iter()
-		.any(|state| state == &issue.state.name)
-	{
-		return ExecutionQueueIntent::NotReady;
-	}
-
-	ExecutionQueueIntent::ReadyToQueue
-}
-
-fn issue_dependencies(
-	issue: &TrackerIssue,
-	supplied_node_ids: &BTreeMap<String, String>,
-) -> Result<Vec<ExecutionProgramDependency>> {
-	let mut dependencies = BTreeMap::new();
-
-	for blocker in &issue.blockers {
-		let dependency_id = supplied_node_ids
-			.get(&blocker.identifier)
-			.cloned()
-			.unwrap_or_else(|| blocker.identifier.clone());
-
-		dependencies
-			.entry(dependency_id.clone())
-			.or_insert(ExecutionProgramDependency::new(dependency_id)?);
-	}
-
-	Ok(dependencies.into_values().collect())
-}
-
-fn dependency_snapshots_for(
-	issue: &TrackerIssue,
-	supplied_node_ids: &BTreeMap<String, String>,
-) -> Result<Vec<ExecutionDependencySnapshot>> {
-	let mut snapshots = BTreeMap::new();
-
-	for blocker in &issue.blockers {
-		let dependency_id = supplied_node_ids
-			.get(&blocker.identifier)
-			.cloned()
-			.unwrap_or_else(|| blocker.identifier.clone());
-		let snapshot = ExecutionDependencySnapshot::tracker_state(
-			dependency_id.clone(),
-			blocker.state.name.clone(),
-		)?;
-
-		snapshots.entry(dependency_id).or_insert(snapshot);
-	}
-
-	Ok(snapshots.into_values().collect())
-}
-
-fn issue_conflict_domains(issue: &TrackerIssue) -> Result<Vec<ExecutionConflictDomain>> {
-	let mut domains = vec![ExecutionConflictDomain::new(
-		ExecutionConflictDomainKind::TrackerOwnership,
-		issue.identifier.clone(),
-	)?];
-	let mut seen = BTreeSet::from([format!(
-		"{}:{}",
-		ExecutionConflictDomainKind::TrackerOwnership.as_str(),
-		issue.identifier
-	)]);
-
-	for label in &issue.labels {
-		if let Some(module) = label.name.strip_prefix("repo:")
-			&& !module.trim().is_empty()
-		{
-			let key = module.trim().to_owned();
-			let seen_key = format!("{}:{key}", ExecutionConflictDomainKind::Module.as_str());
-
-			if seen.insert(seen_key) {
-				domains
-					.push(ExecutionConflictDomain::new(ExecutionConflictDomainKind::Module, key)?);
-			}
-		}
-	}
-
-	domains.sort_by(|left, right| {
-		left.kind().as_str().cmp(right.kind().as_str()).then_with(|| left.key().cmp(right.key()))
-	});
-
-	Ok(domains)
-}
-
-fn issue_report_row(
-	issue: &TrackerIssue,
-	facts: &IssueFacts,
-	evaluation: &ExecutionNodeEvaluation,
-	workflow: &WorkflowDocument,
-) -> IssueBatchIntakeIssueReport {
-	let classification = classify_issue(issue, facts, evaluation, workflow);
-	let mut reasons = evaluation.reasons().to_vec();
-
-	reasons.sort();
-	reasons.dedup();
-
-	let mut blockers =
-		issue.blockers.iter().map(|blocker| blocker.identifier.clone()).collect::<Vec<_>>();
-
-	blockers.sort();
-	blockers.dedup();
-
-	let mut conflict_domains = issue_conflict_domains(issue)
-		.unwrap_or_default()
-		.into_iter()
-		.map(|domain| format!("{}:{}", domain.kind().as_str(), domain.key()))
-		.collect::<Vec<_>>();
-
-	conflict_domains.sort();
-	conflict_domains.dedup();
-	IssueBatchIntakeIssueReport {
-		issue_identifier: issue.identifier.clone(),
-		issue_id: Some(issue.id.clone()),
-		issue_state: Some(issue.state.name.clone()),
-		classification,
-		queue_intent: Some(
-			evaluation
-				.linear_issue()
-				.map_or(ExecutionQueueIntent::NotReady, |_| {
-					issue_queue_intent(issue, facts, workflow)
-				})
-				.as_str()
-				.to_owned(),
-		),
-		dispatch_action: evaluation.dispatch_action().map(dispatch_action_name),
-		reasons,
-		blockers,
-		conflict_domains,
-	}
-}
-
-fn classify_issue(
-	issue: &TrackerIssue,
-	facts: &IssueFacts,
-	evaluation: &ExecutionNodeEvaluation,
-	workflow: &WorkflowDocument,
-) -> IssueBatchIntakeClassification {
-	if state_name_is_terminal(&issue.state.name, workflow) {
-		return IssueBatchIntakeClassification::Stale;
-	}
-	if facts.has_active_label || facts.has_opt_out_label {
-		return IssueBatchIntakeClassification::Held;
-	}
-	if facts.has_needs_attention_label
-		|| facts.has_open_blockers
-		|| !facts.has_generic_dispatch_briefing
-	{
-		return IssueBatchIntakeClassification::Blocked;
-	}
-
-	match evaluation.lifecycle_state() {
-		ExecutionProgramNodeLifecycleState::Ready | ExecutionProgramNodeLifecycleState::Queued =>
-			IssueBatchIntakeClassification::Ready,
-		ExecutionProgramNodeLifecycleState::Planned
-		| ExecutionProgramNodeLifecycleState::Mapped
-		| ExecutionProgramNodeLifecycleState::Active
-		| ExecutionProgramNodeLifecycleState::PostReview => IssueBatchIntakeClassification::Held,
-		ExecutionProgramNodeLifecycleState::Blocked
-		| ExecutionProgramNodeLifecycleState::NeedsAttention => IssueBatchIntakeClassification::Blocked,
-		ExecutionProgramNodeLifecycleState::Completed
-		| ExecutionProgramNodeLifecycleState::Stale
-		| ExecutionProgramNodeLifecycleState::Superseded => IssueBatchIntakeClassification::Stale,
-	}
-}
-
-fn dispatch_action_name(action: ExecutionDispatchAction) -> String {
-	match action {
-		ExecutionDispatchAction::Dispatch => "dispatch",
-	}
-	.to_owned()
-}
-
-fn unmapped_report_row(identifier: &str) -> IssueBatchIntakeIssueReport {
-	IssueBatchIntakeIssueReport {
-		issue_identifier: identifier.to_owned(),
-		issue_id: None,
-		issue_state: None,
-		classification: IssueBatchIntakeClassification::Unmapped,
-		queue_intent: None,
-		dispatch_action: None,
-		reasons: vec![String::from("tracker issue identifier did not resolve")],
-		blockers: Vec::new(),
-		conflict_domains: Vec::new(),
-	}
-}
-
-fn classify_counts(rows: &[IssueBatchIntakeIssueReport]) -> IssueBatchIntakeCounts {
-	let mut counts = IssueBatchIntakeCounts::default();
-
-	for row in rows {
-		match row.classification {
-			IssueBatchIntakeClassification::Ready => counts.ready += 1,
-			IssueBatchIntakeClassification::Held => counts.held += 1,
-			IssueBatchIntakeClassification::Blocked => counts.blocked += 1,
-			IssueBatchIntakeClassification::Stale => counts.stale += 1,
-			IssueBatchIntakeClassification::Unmapped => counts.unmapped += 1,
-		}
-	}
-
-	counts
-}
-
-fn state_name_is_terminal(state_name: &str, workflow: &WorkflowDocument) -> bool {
-	workflow.frontmatter().tracker().terminal_states().iter().any(|state| state == state_name)
-}
-
-fn issue_has_generic_dispatch_briefing(issue: &TrackerIssue) -> bool {
-	orchestrator::issue_has_generic_dispatch_briefing(issue)
-}
-
-#[cfg(test)]
-mod tests {
-	use std::{cell::RefCell, collections::HashMap, fs, path::Path};
-
-	use tempfile::TempDir;
-
-	use crate::{
-		loop_contract::{DecisionContract, DecisionPromotion, DecisionPromotionActorKind},
-		prelude::eyre,
-		program_intake::{
-			self, GoalIntakeIssueAction, GoalIntakeReport, GoalIntakeRunRequest,
-			IssueBatchIntakeClassification,
-		},
-		state::StateStore,
-		tracker::{
-			IssueTracker, TrackerComment, TrackerIssue, TrackerIssueBlocker,
-			TrackerIssueBriefUpdate, TrackerIssueCreate, TrackerLabel, TrackerState, TrackerTeam,
-		},
-		workflow::WorkflowDocument,
-	};
-
-	trait TestIssueExt {
-		fn with_blocker(self, identifier: &str, state: &str) -> Self;
-		fn with_label(self, name: &str) -> Self;
-	}
-
-	#[derive(Default)]
-	struct FakeTracker {
-		issues: RefCell<HashMap<String, TrackerIssue>>,
-		next_issue_number: RefCell<usize>,
-		created_issues: RefCell<Vec<TrackerIssue>>,
-		updated_issues: RefCell<Vec<TrackerIssue>>,
-		fail_create_after_successes: RefCell<Option<usize>>,
-		fail_update_after_successes: RefCell<Option<usize>>,
-	}
-
-	impl FakeTracker {
-		fn with_issues(self, issues: impl IntoIterator<Item = TrackerIssue>) -> Self {
-			for issue in issues {
-				self.issues.borrow_mut().insert(issue.identifier.clone(), issue);
-			}
-
-			self
-		}
-
-		fn with_create_failure_after_successes(self, successes: usize) -> Self {
-			*self.fail_create_after_successes.borrow_mut() = Some(successes);
-
-			self
-		}
-
-		fn with_update_failure_after_successes(self, successes: usize) -> Self {
-			*self.fail_update_after_successes.borrow_mut() = Some(successes);
-
-			self
-		}
-
-		fn created_issue_count(&self) -> usize {
-			self.created_issues.borrow().len()
-		}
-
-		fn updated_issue_count(&self) -> usize {
-			self.updated_issues.borrow().len()
-		}
-
-		fn generated_issue_identifier(&self, index: usize) -> String {
-			format!("XY-G{}", index + 1)
-		}
-	}
-
-	impl IssueTracker for FakeTracker {
-		fn list_issues_with_label(
-			&self,
-			label_name: &str,
-		) -> crate::prelude::Result<Vec<TrackerIssue>> {
-			Ok(self
-				.issues
-				.borrow()
-				.values()
-				.filter(|issue| issue.has_label(label_name))
-				.cloned()
-				.collect())
-		}
-
-		fn find_team_label_id(
-			&self,
-			_team_id: &str,
-			label_name: &str,
-		) -> crate::prelude::Result<Option<String>> {
-			Ok(Some(format!("label-{label_name}")))
-		}
-
-		fn get_issue_by_identifier(
-			&self,
-			issue_identifier: &str,
-		) -> crate::prelude::Result<Option<TrackerIssue>> {
-			Ok(self.issues.borrow().get(issue_identifier).cloned())
-		}
-
-		fn refresh_issues(
-			&self,
-			_issue_ids: &[String],
-		) -> crate::prelude::Result<Vec<TrackerIssue>> {
-			Ok(Vec::new())
-		}
-
-		fn list_comments(&self, _issue_id: &str) -> crate::prelude::Result<Vec<TrackerComment>> {
-			Ok(Vec::new())
-		}
-
-		fn update_issue_state(
-			&self,
-			_issue_id: &str,
-			_state_id: &str,
-		) -> crate::prelude::Result<()> {
-			Ok(())
-		}
-
-		fn add_issue_labels(
-			&self,
-			_issue_id: &str,
-			_label_ids: &[String],
-		) -> crate::prelude::Result<()> {
-			Ok(())
-		}
-
-		fn remove_issue_labels(
-			&self,
-			_issue_id: &str,
-			_label_ids: &[String],
-		) -> crate::prelude::Result<()> {
-			Ok(())
-		}
-
-		fn create_comment(&self, _issue_id: &str, _body: &str) -> crate::prelude::Result<()> {
-			Ok(())
-		}
-
-		fn create_issue(
-			&self,
-			request: &TrackerIssueCreate,
-		) -> crate::prelude::Result<TrackerIssue> {
-			if let Some(success_limit) = *self.fail_create_after_successes.borrow()
-				&& self.created_issues.borrow().len() >= success_limit
-			{
-				eyre::bail!("injected create failure after {success_limit} successes");
-			}
-
-			let identifier = loop {
-				let mut next_issue_number = self.next_issue_number.borrow_mut();
-
-				*next_issue_number += 1;
-
-				let candidate = self.generated_issue_identifier(*next_issue_number - 1);
-
-				if !self.issues.borrow().contains_key(&candidate) {
-					break candidate;
-				}
-			};
-			let state_name = request
-				.state_id
-				.as_deref()
-				.and_then(|state_id| state_id.strip_prefix("state-"))
-				.unwrap_or("Todo");
-			let mut issue = issue(&identifier, state_name);
-
-			issue.id = format!("id-{identifier}");
-
-			issue.title.clone_from(&request.title);
-			issue.description.clone_from(&request.description);
-			issue.team.id.clone_from(&request.team_id);
-			self.issues.borrow_mut().insert(identifier, issue.clone());
-			self.created_issues.borrow_mut().push(issue.clone());
-
-			Ok(issue)
-		}
-
-		fn update_issue_brief(
-			&self,
-			issue_id: &str,
-			request: &TrackerIssueBriefUpdate,
-		) -> crate::prelude::Result<TrackerIssue> {
-			if let Some(success_limit) = *self.fail_update_after_successes.borrow()
-				&& self.updated_issues.borrow().len() >= success_limit
-			{
-				eyre::bail!("injected update failure after {success_limit} successes");
-			}
-
-			let mut issues = self.issues.borrow_mut();
-			let issue = issues
-				.values_mut()
-				.find(|issue| issue.id == issue_id)
-				.ok_or_else(|| eyre::eyre!("issue `{issue_id}` not found"))?;
-
-			issue.title.clone_from(&request.title);
-			issue.description.clone_from(&request.description);
-
-			let issue = issue.clone();
-
-			self.updated_issues.borrow_mut().push(issue.clone());
-
-			Ok(issue)
-		}
-	}
-
-	impl TestIssueExt for TrackerIssue {
-		fn with_blocker(mut self, identifier: &str, state: &str) -> Self {
-			self.blockers.push(TrackerIssueBlocker {
-				id: format!("id-{identifier}"),
-				identifier: identifier.to_owned(),
-				state: TrackerState { id: format!("state-{state}"), name: state.to_owned() },
-			});
-
-			self
-		}
-
-		fn with_label(mut self, name: &str) -> Self {
-			self.labels.push(TrackerLabel { id: format!("label-{name}"), name: name.to_owned() });
-
-			self
-		}
-	}
-
-	#[test]
-	fn issue_batch_dry_run_classifies_without_persisting() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let workflow = workflow();
-		let config = test_config();
-		let tracker = FakeTracker::default().with_issues([
-			issue("XY-1", "Todo"),
-			issue("XY-2", "In Progress"),
-			issue("XY-3", "Done"),
-			issue("XY-4", "Todo")
-				.with_blocker("XY-20", "Todo")
-				.with_blocker("XY-10", "Todo")
-				.with_label("repo:zeta")
-				.with_label("repo:alpha"),
-		]);
-		let report = program_intake::run_issue_batch_intake(
-			&store,
-			&tracker,
-			&config,
-			&workflow,
-			vec![
-				String::from("XY-4"),
-				String::from("XY-2"),
-				String::from("XY-404"),
-				String::from("XY-1"),
-				String::from("XY-3"),
-			],
-			true,
-			false,
-		)
-		.expect("dry-run should classify");
-
-		assert_eq!(report.counts.ready, 1);
-		assert_eq!(report.counts.held, 1);
-		assert_eq!(report.counts.blocked, 1);
-		assert_eq!(report.counts.stale, 1);
-		assert_eq!(report.counts.unmapped, 1);
-		assert_eq!(report.issues[0].issue_identifier, "XY-1");
-		assert_eq!(report.issues[0].classification, IssueBatchIntakeClassification::Ready);
-
-		let blocked = report
-			.issues
-			.iter()
-			.find(|issue| issue.issue_identifier == "XY-4")
-			.expect("blocked issue should be reported");
-
-		assert_eq!(blocked.blockers, vec![String::from("XY-10"), String::from("XY-20")]);
-		assert_eq!(
-			blocked.conflict_domains,
-			vec![
-				String::from("module:alpha"),
-				String::from("module:zeta"),
-				String::from("tracker_ownership:XY-4"),
-			]
-		);
-		assert!(
-			store.list_execution_programs("decodex").expect("program list should read").is_empty()
-		);
-	}
-
-	#[test]
-	fn project_registration_is_persist_only_for_command_path() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let temp_dir = TempDir::new().expect("temp dir should create");
-		let config_path = write_project_files(temp_dir.path());
-
-		program_intake::register_intake_project_config_for_persist(&store, &config_path, false)
-			.expect("dry-run registration should no-op");
-
-		assert!(store.list_projects().expect("projects should list").is_empty());
-
-		program_intake::register_intake_project_config_for_persist(&store, &config_path, true)
-			.expect("persist registration should write");
-
-		let projects = store.list_projects().expect("projects should list");
-
-		assert_eq!(projects.len(), 1);
-		assert_eq!(projects[0].service_id(), "decodex");
-		assert!(projects[0].enabled());
-	}
-
-	#[test]
-	fn issue_batch_persist_writes_program_and_adjacent_intake_state() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let workflow = workflow();
-		let config = test_config();
-		let tracker = FakeTracker::default().with_issues([issue("XY-1", "Todo")]);
-		let report = program_intake::run_issue_batch_intake(
-			&store,
-			&tracker,
-			&config,
-			&workflow,
-			vec![String::from("XY-1")],
-			false,
-			true,
-		)
-		.expect("persist should write local state");
-
-		assert!(report.persisted);
-		assert_eq!(store.list_execution_programs("decodex").expect("programs").len(), 1);
-		assert_eq!(store.list_program_intake_plans("decodex").expect("plans").len(), 1);
-		assert_eq!(
-			store
-				.list_program_issue_mappings("decodex", &report.program_id)
-				.expect("mappings")
-				.len(),
-			1
-		);
-		assert_eq!(
-			store.list_program_intake_plans("decodex").expect("plans")[0].intake_kind(),
-			"issue_batch_intake"
-		);
-	}
-
-	#[test]
-	fn goal_intake_dry_run_shows_issue_split_without_mutation() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let contract = accepted_goal_contract();
-
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), contract)
-			.expect("contract should persist");
-
-		let tracker = FakeTracker::default().with_issues([issue("XY-852", "Todo")]);
-		let config = test_config();
-		let workflow = workflow();
-		let report = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: true,
-			apply: false,
-		})
-		.expect("dry-run should produce materialization plan");
-
-		assert!(report.dry_run);
-		assert!(!report.persisted);
-		assert_eq!(report.issues.len(), 2);
-		assert_eq!(report.issues[0].action, GoalIntakeIssueAction::WouldCreate);
-		assert_eq!(report.issues[0].dependencies, Vec::<String>::new());
-		assert_eq!(report.issues[0].conflict_domains, vec![String::from("module:runtime")]);
-		assert_eq!(report.issues[1].dependencies, vec![String::from("goal-intake-runtime")]);
-		assert_eq!(tracker.created_issue_count(), 0);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-
-		let rendered = program_intake::render_goal_intake_report(&report);
-
-		assert!(rendered.contains("dependencies=none"));
-		assert!(rendered.contains("conflict_domains=module:runtime"));
-		assert!(rendered.contains("dependencies=goal-intake-runtime"));
-	}
-
-	#[test]
-	fn goal_intake_refuses_latent_or_missing_decision_authority() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let tracker = FakeTracker::default().with_issues([issue("XY-852", "Todo")]);
-		let latent = latent_goal_contract();
-
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), latent)
-			.expect("latent contract should persist");
-
-		let config = test_config();
-		let workflow = workflow();
-		let latent_error = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: false,
-			apply: true,
-		})
-		.expect_err("latent contract must not materialize");
-
-		assert!(latent_error.to_string().contains("requires accepted execution authority"));
-
-		let mut needs_decision = latent_goal_contract();
-
-		needs_decision
-			.require_human_decision("Choose the public issue split before apply.")
-			.expect("contract should record missing decision");
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), needs_decision)
-			.expect("needs-decision contract should persist");
-
-		let missing_decision_error = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: false,
-			apply: true,
-		})
-		.expect_err("missing decision must stop apply");
-
-		assert!(missing_decision_error.to_string().contains("needs_human_decision"));
-		assert_eq!(tracker.created_issue_count(), 0);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-	}
-
-	#[test]
-	fn goal_intake_apply_creates_updates_and_persists_links() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let mut contract = accepted_goal_contract();
-
-		contract
-			.link_generated_execution_surfaces(["id-XY-G1"], ["XY-G1"], ["old-node"])
-			.expect("existing generated link should attach");
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), contract)
-			.expect("contract should persist");
-
-		let tracker =
-			FakeTracker::default().with_issues([issue("XY-852", "Todo"), issue("XY-G1", "Todo")]);
-		let config = test_config();
-		let workflow = workflow();
-		let report = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: false,
-			apply: true,
-		})
-		.expect("apply should materialize issues and program");
-
-		assert_goal_intake_apply_report(&report, &tracker);
-		assert_goal_intake_runtime_links(&store, &report);
-
-		let updated = tracker
-			.get_issue_by_identifier("XY-G1")
-			.expect("issue lookup should work")
-			.expect("updated issue should exist");
-
-		assert_goal_issue_brief_is_public(&updated.description, &report);
-	}
-
-	fn assert_goal_intake_apply_report(report: &GoalIntakeReport, tracker: &FakeTracker) {
-		assert!(report.applied);
-		assert!(report.persisted);
-		assert_eq!(tracker.updated_issue_count(), 1);
-		assert_eq!(tracker.created_issue_count(), 1);
-		assert_eq!(report.issues[0].action, GoalIntakeIssueAction::Updated);
-		assert_eq!(report.issues[1].action, GoalIntakeIssueAction::Created);
-		assert_eq!(report.issues[0].dispatch_action.as_deref(), Some("dispatch"));
-		assert_eq!(report.issues[1].dispatch_action.as_deref(), None);
-		assert!(
-			report.issues[1]
-				.reasons
-				.iter()
-				.any(|reason| reason.contains("has not reached a required terminal state"))
-		);
-	}
-
-	fn assert_goal_intake_runtime_links(store: &StateStore, report: &GoalIntakeReport) {
-		let linked_contract = store
-			.decision_contract("decodex", "goal-intake-contract")
-			.expect("contract lookup should read")
-			.expect("contract should exist");
-		let node_ids = report.issues.iter().map(|issue| issue.node_id.clone()).collect::<Vec<_>>();
-
-		assert_eq!(
-			linked_contract.contract().links().generated_issue_identifiers(),
-			&[String::from("XY-G1"), String::from("XY-G2")]
-		);
-		assert_eq!(linked_contract.contract().links().execution_program_node_ids(), &node_ids);
-
-		let programs = store
-			.list_execution_programs_for_contract("decodex", "goal-intake-contract")
-			.expect("programs should list");
-		let intake_plans =
-			store.list_program_intake_plans("decodex").expect("intake plans should list");
-
-		assert_eq!(programs.len(), 1);
-		assert_eq!(programs[0].program_id(), report.program_id);
-		assert_eq!(intake_plans.len(), 1);
-		assert_eq!(intake_plans[0].intake_kind(), "goal_intake");
-		assert_eq!(intake_plans[0].source_contract_id(), Some("goal-intake-contract"));
-
-		assert_goal_intake_plan_lineage(&programs[0]);
-
-		assert_eq!(
-			store
-				.list_program_issue_mappings("decodex", &report.program_id)
-				.expect("mappings should list")
-				.len(),
-			2
-		);
-	}
-
-	fn assert_goal_intake_plan_lineage(record: &crate::state::ExecutionProgramRecord) {
-		let intake_plan = record
-			.program()
-			.program_intake_plan()
-			.expect("program payload should retain intake plan lineage");
-
-		assert_eq!(intake_plan.source_objective_ref(), Some("decodex:quality-autonomy@1"));
-		assert_eq!(intake_plan.source_proposal_id(), Some("autonomy_proposal:test-proposal"));
-		assert_eq!(
-			intake_plan.source_signal_refs(),
-			&[String::from("autonomy_signal:test-signal")]
-		);
-	}
-
-	fn assert_goal_issue_brief_is_public(description: &str, report: &GoalIntakeReport) {
-		for heading in [
-			"## Objective",
-			"## Authority",
-			"## Required Reading",
-			"## Ownership Boundary",
-			"## Dependencies",
-			"## Current-tree Landing Zone",
-			"## Acceptance",
-			"## Validation",
-			"## Lifecycle Gates",
-			"## Risk",
-			"## Stop Conditions",
-		] {
-			assert!(description.contains(heading));
-		}
-
-		assert!(description.contains(
-			"Accepted Decision Contract authority is recorded in Decodex runtime state."
-		));
-		assert!(description.contains("Source issue: `XY-852`"));
-		assert!(description.contains("Goal intake dry-run renders generated issue briefs"));
-		assert!(description.contains("Use normal Decodex review, PR handoff, landing"));
-		assert!(description.contains("Run install or restart steps only when"));
-		assert!(description.contains("Stop when promotion authority"));
-
-		assert_goal_issue_brief_hides_private_ids(description, report);
-	}
-
-	fn assert_goal_issue_brief_hides_private_ids(description: &str, report: &GoalIntakeReport) {
-		for private_id in [
-			"Execution Program: `",
-			"Execution Program node:",
-			"goal-intake-contract",
-			"autonomy_proposal:test-proposal",
-			"decodex:quality-autonomy@1",
-			"autonomy_signal:test-signal",
-			&report.program_id,
-			"```",
-			"private_evidence_refs",
-		] {
-			assert!(!description.contains(private_id));
-		}
-		for issue in &report.issues {
-			assert!(!description.contains(&issue.node_id));
-		}
-	}
-
-	#[test]
-	fn generated_issue_text_validation_rejects_private_program_identifiers() {
-		let title_error = program_intake::validate_generated_issue_text(
-			"Expose goal-decodex-contract-private",
-			"## Objective\nUse normal public text.",
-			&["goal-decodex-contract-private"],
-		)
-		.expect_err("title must reject private program id");
-
-		assert!(
-			title_error
-				.to_string()
-				.contains("generated issue title contains a private Program Intake identifier")
-		);
-
-		let description_error = program_intake::validate_generated_issue_text(
-			"Use normal public text.",
-			"## Objective\nExpose goal:contract:01-private-node.",
-			&["goal:contract:01-private-node"],
-		)
-		.expect_err("description must reject private node id");
-
-		assert!(
-			description_error.to_string().contains(
-				"generated issue description contains a private Program Intake identifier"
-			)
-		);
-	}
-
-	#[test]
-	fn goal_intake_apply_rejects_generated_briefs_that_leak_autonomy_lineage_ids() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let mut payload = serde_json::to_value(accepted_goal_contract())
-			.expect("accepted goal contract should serialize");
-
-		payload["accepted_authority"]["constraints"]
-			.as_array_mut()
-			.expect("constraints should be an array")
-			.push(serde_json::json!(
-				"Do not expose autonomy_signal:test-signal in generated issue text."
-			));
-
-		let leaking_contract: DecisionContract =
-			serde_json::from_value(payload).expect("leaking contract should deserialize");
-
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), leaking_contract)
-			.expect("contract should persist");
-
-		let tracker = FakeTracker::default().with_issues([issue("XY-852", "Todo")]);
-		let config = test_config();
-		let workflow = workflow();
-		let error = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: false,
-			apply: true,
-		})
-		.expect_err("private autonomy lineage ids must fail generated brief validation");
-
-		assert!(
-			error.to_string().contains(
-				"generated issue description contains a private Program Intake identifier"
-			)
-		);
-		assert_eq!(tracker.created_issue_count(), 0);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-	}
-
-	#[test]
-	fn goal_intake_apply_persists_links_after_each_successful_issue_mutation() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let contract = accepted_goal_contract();
-
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), contract)
-			.expect("contract should persist");
-
-		let tracker = FakeTracker::default()
-			.with_issues([issue("XY-852", "Todo")])
-			.with_create_failure_after_successes(1);
-		let config = test_config();
-		let workflow = workflow();
-		let error = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: false,
-			apply: true,
-		})
-		.expect_err("second issue create should fail");
-
-		assert!(error.to_string().contains("injected create failure"));
-		assert_eq!(tracker.created_issue_count(), 1);
-
-		let linked_contract = store
-			.decision_contract("decodex", "goal-intake-contract")
-			.expect("contract lookup should read")
-			.expect("contract should exist");
-
-		assert_eq!(
-			linked_contract.contract().links().generated_issue_identifiers(),
-			&[String::from("XY-G1")]
-		);
-		assert_eq!(linked_contract.contract().links().generated_issue_ids().len(), 1);
-		assert_eq!(linked_contract.contract().links().execution_program_node_ids().len(), 1);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-	}
-
-	#[test]
-	fn goal_intake_apply_preserves_later_existing_links_after_update_failure() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let mut contract = accepted_goal_contract();
-
-		contract
-			.link_generated_execution_surfaces(
-				["id-XY-G1", "id-XY-G2"],
-				["XY-G1", "XY-G2"],
-				["old-node-1", "old-node-2"],
-			)
-			.expect("existing generated links should attach");
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), contract)
-			.expect("contract should persist");
-
-		let tracker = FakeTracker::default()
-			.with_issues([issue("XY-852", "Todo"), issue("XY-G1", "Todo"), issue("XY-G2", "Todo")])
-			.with_update_failure_after_successes(1);
-		let config = test_config();
-		let workflow = workflow();
-		let error = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: false,
-			apply: true,
-		})
-		.expect_err("second issue update should fail");
-
-		assert!(error.to_string().contains("injected update failure"));
-		assert_eq!(tracker.updated_issue_count(), 1);
-
-		let linked_contract = store
-			.decision_contract("decodex", "goal-intake-contract")
-			.expect("contract lookup should read")
-			.expect("contract should exist");
-
-		assert_eq!(
-			linked_contract.contract().links().generated_issue_identifiers(),
-			&[String::from("XY-G1"), String::from("XY-G2")]
-		);
-		assert_eq!(linked_contract.contract().links().execution_program_node_ids().len(), 2);
-		assert_eq!(
-			linked_contract.contract().links().execution_program_node_ids()[1],
-			"old-node-2"
-		);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-	}
-
-	#[test]
-	fn goal_intake_apply_fails_closed_when_existing_generated_link_is_missing() {
-		let store = StateStore::open_in_memory().expect("store should open");
-		let mut contract = accepted_goal_contract();
-
-		contract
-			.link_generated_execution_surfaces(["id-XY-G1"], ["XY-G1"], ["old-node"])
-			.expect("existing generated link should attach");
-		store
-			.upsert_decision_contract("decodex", Some("XY-852"), contract)
-			.expect("contract should persist");
-
-		let tracker = FakeTracker::default().with_issues([issue("XY-852", "Todo")]);
-		let config = test_config();
-		let workflow = workflow();
-		let error = program_intake::run_goal_intake(GoalIntakeRunRequest {
-			state_store: &store,
-			tracker: &tracker,
-			config: &config,
-			workflow: &workflow,
-			contract_id: "goal-intake-contract",
-			team_issue_identifier: None,
-			dry_run: false,
-			apply: true,
-		})
-		.expect_err("missing generated issue link should block apply");
-
-		assert!(error.to_string().contains("Generated issue link `XY-G1`"));
-		assert_eq!(tracker.created_issue_count(), 0);
-		assert_eq!(tracker.updated_issue_count(), 0);
-		assert!(store.list_execution_programs("decodex").expect("programs").is_empty());
-	}
-
-	fn workflow() -> WorkflowDocument {
-		WorkflowDocument::parse_markdown(workflow_markdown()).expect("workflow should parse")
-	}
-
-	fn latent_goal_contract() -> DecisionContract {
-		serde_json::from_value(latent_goal_contract_payload())
-			.expect("goal contract should deserialize")
-	}
-
-	fn latent_goal_contract_payload() -> serde_json::Value {
-		serde_json::json!({
-			"schema": crate::loop_contract::DECISION_CONTRACT_SCHEMA,
-			"record_version": crate::loop_contract::DECISION_CONTRACT_RECORD_VERSION,
-			"contract_id": "goal-intake-contract",
-			"status": "draft_latent",
-			"source_intent": latent_goal_source_intent(),
-			"research_provenance": latent_goal_research_provenance(),
-			"research_evidence": latent_goal_research_evidence(),
-			"research_options": [],
-			"accepted_authority": latent_goal_accepted_authority(),
-			"execution_readiness": latent_goal_execution_readiness(),
-			"links": {
-				"generated_issue_ids": [],
-				"generated_issue_identifiers": [],
-			"execution_program_node_ids": []
-		},
-		"evidence_boundary": {
-			"private_evidence_refs": [],
-			"public_projection_refs": [],
-				"public_summary": "Goal intake contract ready for issue shaping."
-			}
-		})
-	}
-
-	fn latent_goal_source_intent() -> serde_json::Value {
-		serde_json::json!({
-			"summary": "Ship promoted goal intake.",
-			"user_utterance": "arrange this goal",
-			"source_issue_identifier": "XY-852",
-		})
-	}
-
-	fn latent_goal_research_provenance() -> serde_json::Value {
-		serde_json::json!([
-			{
-				"kind": "autonomy_proposal",
-				"reference": "autonomy_proposal:test-proposal",
-				"summary": "Accepted autonomy proposal produced this Decision Contract candidate."
-			},
-			{
-				"kind": "autonomy_objective",
-				"reference": "decodex:quality-autonomy@1",
-				"summary": "Accepted autonomy objective version."
-			},
-			{
-				"kind": "spec",
-				"reference": "docs/spec/loop-runtime.md",
-				"summary": "Promoted contracts can shape normal Linear issues."
-			}
-		])
-	}
-
-	fn latent_goal_research_evidence() -> serde_json::Value {
-		serde_json::json!([
-			{
-				"kind": "autonomy_signal:runtime_health",
-				"claim": "Autonomy signal `autonomy_signal:test-signal` contributed.",
-				"support": "freshness=fresh; evidence_class=repo_source; confidence=high",
-				"source_ref": "autonomy_signal:test-signal"
-			},
-			{
-				"claim": "Goal intake needs generated issues and an internal program.",
-				"support": "The loop-runtime spec defines Program Intake records.",
-				"source_ref": "docs/spec/loop-runtime.md"
-			}
-		])
-	}
-
-	fn latent_goal_accepted_authority() -> serde_json::Value {
-		serde_json::json!({
-			"accepted_objectives": [
-				"Materialize accepted goal intake into normal Linear issues.",
-				"Persist the internal Execution Program without exposing graph mechanics."
-			],
-			"non_goals": ["Do not run implementation from goal intake."],
-			"constraints": ["Linear receives only public-safe issue briefs and sparse links."],
-			"assumptions": ["The source issue anchors the generated issue team."],
-			"objections": [],
-			"stop_conditions": [
-				"Stop when promotion authority or required decisions are missing."
-			]
-		})
-	}
-
-	fn latent_goal_execution_readiness() -> serde_json::Value {
-		serde_json::json!({
-			"summary": "Ready for issue shaping after promotion.",
-			"ready_for_issue_shaping": true,
-			"missing_decisions": [],
-			"validation_expectations": ["Run cargo make test before handoff."],
-			"risk_notes": ["Generated issue descriptions must stay natural-language."],
-			"proposed_issues": [goal_intake_runtime_issue(), goal_intake_links_issue()],
-			"conflict_domains": ["module:runtime", "file:docs/spec/loop-runtime.md"]
-		})
-	}
-
-	fn goal_intake_runtime_issue() -> serde_json::Value {
-		serde_json::json!({
-			"key": "goal-intake-runtime",
-			"title": "Implement goal intake CLI/API behavior.",
-			"objective": "Implement goal intake CLI/API behavior.",
-			"stage": "runtime",
-			"dependencies": [],
-			"conflict_domains": ["module:runtime"],
-			"acceptance": ["Goal intake dry-run renders generated issue briefs without mutation."],
-			"validation": ["Run cargo make test before handoff."],
-			"risk": ["Generated issue descriptions must stay natural-language."],
-			"queue_intent": "ready_to_queue"
-		})
-	}
-
-	fn goal_intake_links_issue() -> serde_json::Value {
-		serde_json::json!({
-			"key": "goal-intake-links",
-			"title": "Persist Execution Program links for generated issues.",
-			"objective": "Persist Execution Program links for generated issues.",
-			"stage": "runtime",
-			"dependencies": ["goal-intake-runtime"],
-			"conflict_domains": ["module:runtime", "file:docs/spec/loop-runtime.md"],
-			"acceptance": [
-				"Apply links generated issue identifiers and execution nodes back to the accepted contract."
-			],
-			"validation": ["Run cargo make test before handoff."],
-			"risk": ["Generated issue descriptions must stay natural-language."],
-			"queue_intent": "ready_to_queue"
-		})
-	}
-
-	fn accepted_goal_contract() -> DecisionContract {
-		let mut contract = latent_goal_contract();
-
-		contract
-			.promote(
-				DecisionPromotion::new(
-					"operator",
-					DecisionPromotionActorKind::User,
-					"2026-06-10T00:00:00Z",
-					"conversation",
-					Some(String::from("User asked Decodex to arrange this goal.")),
-				)
-				.expect("promotion should build"),
-			)
-			.expect("contract should promote");
-
-		contract
-	}
-
-	fn workflow_markdown() -> &'static str {
-		r#"+++
-version = 1
-[tracker]
-provider = "linear"
-startable_states = ["Todo"]
-terminal_states = ["Done", "Canceled", "Duplicate"]
-in_progress_state = "In Progress"
-success_state = "In Review"
-completed_state = "Done"
-failure_state = "Todo"
-opt_out_label = "decodex:manual-only"
-needs_attention_label = "decodex:needs-attention"
-[agent]
-transport = "stdio://"
-[execution]
-max_attempts = 3
-max_turns = 3
-max_retry_backoff_ms = 300000
-gate_profiles = {}
-canonicalize_commands = ["cargo make fmt"]
-verify_commands = ["cargo make test"]
-[execution.workspace_hooks]
-after_create_commands = []
-before_remove_commands = []
-timeout_seconds = 60
-[context]
-read_first = []
-+++
-"#
-	}
-
-	fn test_config() -> crate::config::ServiceConfig {
-		crate::config::ServiceConfig::parse_toml(
-			r#"
-service_id = "decodex"
-[tracker]
-api_key_env_var = "HOME"
-[github]
-token_env_var = "HOME"
-[codex]
-review = "standard"
-[paths]
-repo_root = "."
-worktree_root = ".worktrees"
-"#,
-		)
-		.expect("config should parse")
-	}
-
-	fn write_project_files(project_dir: &Path) -> std::path::PathBuf {
-		fs::write(project_dir.join("WORKFLOW.md"), workflow_markdown())
-			.expect("workflow should write");
-		fs::write(
-			project_dir.join("project.toml"),
-			r#"
-service_id = "decodex"
-[tracker]
-api_key_env_var = "HOME"
-[github]
-token_env_var = "HOME"
-[codex]
-review = "standard"
-[paths]
-repo_root = "."
-worktree_root = ".worktrees"
-"#,
-		)
-		.expect("project config should write");
-
-		project_dir.join("project.toml")
-	}
-
-	fn issue(identifier: &str, state: &str) -> TrackerIssue {
-		TrackerIssue {
-			id: format!("id-{identifier}"),
-			identifier: identifier.to_owned(),
-			project_slug: None,
-			title: format!("Issue {identifier}"),
-			author: None,
-			description: format!("Implement {identifier}."),
-			priority: None,
-			created_at: String::from("2026-06-01T00:00:00Z"),
-			updated_at: String::from("2026-06-01T00:00:00Z"),
-			state: TrackerState { id: format!("state-{state}"), name: state.to_owned() },
-			team: TrackerTeam {
-				id: String::from("team"),
-				name: String::from("Team"),
-				states: vec![
-					TrackerState { id: String::from("state-Todo"), name: String::from("Todo") },
-					TrackerState {
-						id: String::from("state-In Progress"),
-						name: String::from("In Progress"),
-					},
-					TrackerState { id: String::from("state-Done"), name: String::from("Done") },
-				],
-				labels: Vec::new(),
-			},
-			labels_complete: true,
-			labels: Vec::new(),
-			blockers: Vec::new(),
-		}
-	}
 }
